==> analysis/00_exploreCpGGST.Rmd <==
---
title: "00_exploreCpGGST"
author: "Caitlin Page"
date: "2025-03-05"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: inline
---

## Packages

```{r}
library(org.Hs.eg.db)
library(GO.db)

library(plyranges)
library(dplyr)
library(ggplot2)
```

To do the Gene Set Test, we need to obtain all the gene sets, and the genes that are in them.
The trick is, that this can include the child terms.
The other trick is that we don't want this to be species dependent.
And although we have mainly been using GO, we want it to applicable for other gene set databases like KEGG.

## Previous work

```{r}
# missMethyl
.getGO <- function(){
  if(!requireNamespace("org.Hs.eg.db", quietly = TRUE))
    stop("org.Hs.eg.db package required but not installed.")
  egGO2ALLEGS <- utils::getFromNamespace("org.Hs.egGO2ALLEGS", "org.Hs.eg.db")
  GeneID.PathID <- AnnotationDbi::toTable(egGO2ALLEGS)[,c("gene_id", "go_id", "Ontology")]
  d <- !duplicated(GeneID.PathID[, c("gene_id", "go_id")])
  GeneID.PathID <- GeneID.PathID[d, ]
  GOID.TERM <- suppressMessages(AnnotationDbi::select(GO.db::GO.db,
                                                      keys=unique(GeneID.PathID$go_id),
                                                      columns=c("GOID","ONTOLOGY","TERM"),
                                                      keytype="GOID"))
  go <- tapply(GeneID.PathID$gene_id, GeneID.PathID$go_id, list)

  list(idList=go, idTable=GOID.TERM)
}
```
* Positives: doesn't require an input - just runs
             easy to read
* Negatives: only allows Human input - not other species

```{r}
# goseq

getgo=function(genes,genome,id,fetch.cats=c("GO:CC","GO:BP","GO:MF")){
	#Check for valid input
	if(any(!fetch.cats%in%c("GO:CC","GO:BP","GO:MF","KEGG"))){
		stop("Invaled category specified.  Categories can only be GO:CC, GO:BP, GO:MF or KEGG")
	}
	#Convert from genome ID to org.__.__.db format
	orgstring=as.character(.ORG_PACKAGES[match(gsub("[0-9]+",'',genome),names(.ORG_PACKAGES))])
	#Multimatch or no match
	if(length(orgstring)!=1){
		stop("Couldn't grab GO categories automatically.  Please manually specify.")
	}
	#Load the library
	library(paste(orgstring,"db",sep='.'),character.only=TRUE)
	#What is the default ID that the organism package uses?
	coreid=strsplit(orgstring,"\\.")[[1]][3]

	#Now we need to convert it into the naming convention used by the organism packages
	userid=as.character(.ID_MAP[match(id,names(.ID_MAP))])
	#Multimatch or no match
	if(is.na(userid) | (length(userid)!=1)){
		stop("Couldn't grab GO categories automatically.  Please manually specify.")
	}
	#The (now loaded) organism package contains a mapping between the internal ID and whatever 
        #the default is (usually eg), the rest of this function is about changing that mapping to 
	#point from categories to the ID specified
	#Fetch the mapping in its current format
	#Because GO is a directed graph, we need to get not just the genes associated with each ID, 
	#but also those associated with its children.  GO2ALLEGS does this.
	core2cat=NULL
	if(length(grep("^GO",fetch.cats))!=0){
	        #Get the name of the function which maps gene ids to go terms
		#usually this will be "GO2ALLEG"
	        gomapFunction=.ORG_GOMAP_FUNCTION[orgstring]
		if(is.na(gomapFunction)) gomapFunction=.ORG_GOMAP_FUNCTION["default"]	
		x=toTable(get(paste(orgstring,gomapFunction,sep='')))
		#Keep only those ones that we specified and keep only the names
#		core2cat=x[x$Ontology%in%gsub("^GO:",'',fetch.cats),1:2]
		x[!x$Ontology%in%gsub("^GO:",'',fetch.cats),2]<-"Other"
		core2cat=x[,1:2]
		colnames(core2cat)=c("gene_id","category")
	}
	if(length(grep("^KEGG",fetch.cats))!=0){
		x=toTable(get(paste(orgstring,"PATH",sep='')))
		#Either add it to existing table or create a new one
		colnames(x)=c("gene_id","category")
		if(!is.null(core2cat)){
			core2cat=rbind(core2cat,x)
		}else{
			core2cat=x
		}
	}

	#Now we MAY have to convert the "gene_id" column to the format we are using
	if(coreid!=userid){
		#The mapping between user id and core id, don't use the <USER_ID>2<CORE_ID> object as the naming is not always consistent
		user2core=toTable(get(paste(orgstring,userid,sep='')))
		#Throw away any user ID that doesn't appear in core2cat
		user2core=user2core[user2core[,1]%in%core2cat[,1],]
		#Make a list version of core2cat, we'll need it
		list_core2cat=split(core2cat[,2],core2cat[,1])
		#Now we need to replicate the core IDs that need replicating
		list_core2cat=list_core2cat[match(user2core[,1],names(list_core2cat))]
		#Now we can replace the labels on this list with the user ones from user2core, 
		#but there will be duplicates, so we have to unlist, label, then relist
		user2cat=split(unlist(list_core2cat,FALSE,FALSE),rep(user2core[,2],sapply(list_core2cat,length)))
		#Now we only want each category listed once for each entry...
		user2cat=sapply(user2cat,unique)
		###In case you don't believe that this works as it should, here is the slow as all hell way for comparison...
		###Make first list
		##list_user2core=split(user2core[,1],user2core[,2])
		###Make the second
		##list_core2cat=split(core2cat[,2],core2cat[,1])
		###Go through each entry in first list and expand using second...
		##user2cat=sapply(list_user2core,function(u){unique(unlist(list_core2cat[u],FALSE,FALSE))})

	}else{
		#We don't need to convert anything (WOO!), so just make it into a list
		user2cat=split(core2cat[,2],core2cat[,1])
		user2cat=sapply(user2cat,unique)
	}
	#remove any empty strings
	user2cat=lapply(user2cat,function(x){  
	        if(length(x)>1) x=x[x!="Other"]  
		x })

	## we don't like case sensitivity
	names(user2cat)<-toupper(names(user2cat))

	#Now look them up
	return(user2cat[toupper(genes)])
}
```
* Positives: allows for multiple species
             GO or KEGG in one fn
* Negatives: code is old and not easy to follow

## Developing my way

* Requirements:
** allow for multiple species
** separate methods for GO/KEGG (start with GO)

* Get the GO terms
Use an [org object](10.18129/B9.bioc.org.Hs.eg.db)
These are species dependent, and provides genome wide mapping, using Entrez Gene identifiers
```{r}
org.Hs.eg.db
```

There are 3 objects available for GO ontology terms.
```{r}
c(org.Hs.egGO, org.Hs.egGO2ALLEGS, org.Hs.egGO2EG)
```
[1] maps entrez gene identifiers and GO ids they are directly associated with, NO CHILD TERMS
[2] maps between GO and all of entrez gene ids annotated at that GO or to it's child terms
[3] I think this one is just the reverse mapping of [1]

Which to use?
* missMethyl uses GO2ALLEGS
* unclear what goseq defaults to

AIM:
* explore the impact of using GO or GO2ALLEGS
```{r}
xx <- as.list(org.Hs.egGO2EG)
go_entrez_direct <- data.frame(cbind(go = gsub("\\..*", "", names(unlist(xx))), entrez = unlist(xx)))
go_entrez_direct <- go_entrez_direct %>% group_by(go) %>% mutate(num_genes = n()) %>% ungroup()
go_entrez_direct <- data.frame(go_entrez_direct)
xx <- as.list(org.Hs.egGO2ALLEGS)
go_entrez_allegs <- data.frame(cbind(go = gsub("\\..*", "", names(unlist(xx))), entrez = unlist(xx)))
go_entrez_allegs <- go_entrez_allegs %>% group_by(go) %>% mutate(num_genes = n()) %>% ungroup()
go_entrez_allegs <- data.frame(go_entrez_allegs)
```

```{r}
go_entrez_direct %>% distinct(go)
go_entrez_allegs %>% distinct(go)
```
```{r}
mm_go <- .getGO()
mm_go <- mm_go$idTable
```

```{r}
unique(go_entrez_direct$go) %in% mm_go$GOID %>% summary()
unique(go_entrez_allegs$go) %in% mm_go$GOID %>% summary()
```
```{r}
mm_go <- mm_go %>% mutate(n_genes_direct = go_entrez_direct[match(.$GOID, go_entrez_direct$go), "num_genes"],
                          n_genes_alleg = go_entrez_allegs[match(.$GOID, go_entrez_allegs$go), "num_genes"])
mm_go[is.na(mm_go)] <- 0
head(mm_go)
```
```{r}
mm_go %>%
  ggplot(aes(x = n_genes_direct, y = n_genes_alleg)) +
  geom_point(alpha = 0.3)

mm_go %>%
  ggplot(aes(x = log2(n_genes_direct), y = log2(n_genes_alleg))) +
  geom_point(alpha = 0.3)
```

```{r}
cor.test(mm_go$n_genes_direct, mm_go$n_genes_alleg)
cor.test(mm_go$n_genes_direct, mm_go$n_genes_alleg, method = "kendall")
cor.test(mm_go$n_genes_direct, mm_go$n_genes_alleg, method = "spearman")
```

* number of genes is not very correlated
* parametric probably not the best way to do the test because the data does not follow the normal distribution
* Spearman gives correlation of 0.5 (so semi correlated)

```{r}
mm_go %>% filter(n_genes_direct == 0) %>% nrow(.)
```

```{r}
mm_go %>% mutate(direct_0 = ifelse(n_genes_direct == 0, TRUE, FALSE)) %>%
  ggplot(aes(x = log2(n_genes_alleg), colour = direct_0)) +
  geom_density()

mm_go %>% mutate(direct_0 = ifelse(n_genes_direct == 0, TRUE, FALSE)) %>%
  ggplot(aes(x = direct_0, y = log2(n_genes_alleg), fill = direct_0)) +
  geom_boxplot()
```
* gene sets with no genes in direct do not appear to be a subgroup of the allegs 

* it would be nice to have a hierachy map with all of this to see how the gene sets are related
(I know I looked into a few things back in the day, see if I can find them)
* add in some cpg bias stuff here
* but make the gene stuff a separate file so it doesn't get too chaotic
* also add in final decision I guess: so run the method with both? and see how results change?
* or separate again?
Analysis continues in converting gene ids [here](03_genesToGST.html)

## CpG Bias

* oh except this requires the correct gene swaps
* or just the results overall?

## the gene stuff - we're putting it back lol

```{r}

xx <- as.list(org.Hs.egGO2ALLEGS)

go_entrez <- data.frame(cbind(go = gsub("\\..*", "", names(unlist(xx))), entrez = unlist(xx)))

# map entrez to gene symbol https://web.mit.edu/~r/current/arch/i386_linux26/lib/R/library/org.Hs.eg.db/html/org.Hs.egSYMBOL.html
x <- org.Hs.egSYMBOL
# Get the gene symbol that are mapped to an entrez gene identifiers
mapped_genes <- mappedkeys(x)
# Convert to a list
xx <- as.list(x[mapped_genes])

entrez_symbol <- data.frame(cbind(entrez = gsub("\\..*", "", names(unlist(xx))), symbol = unlist(xx)))
go_entrez <- go_entrez %>% mutate(symbol = entrez_symbol[match(.$entrez, entrez_symbol$entrez), "symbol"])
go_entrez <- go_entrez %>% mutate(ontology = mm_go[match(.$go, mm_go$GOID), "ONTOLOGY"],
                                  term = mm_go[match(.$go, mm_go$GOID), "TERM"])
go_entrez %>% filter(entrez == 1)
```
* theoretically can get positions for org package but I think it's icky
* and no idea what genome it would follow
* do a thing for txdb
* but right now just carry over biomart
https://support.bioconductor.org/p/p133188/
```{r}
biomart_genes <- readRDS("../output/biomart_genes.rds")
biomart_genes <- biomart_genes %>% mutate(coords = paste0(seqnames, ":", start, "-", end))
biomart_genes <- data.frame(biomart_genes)
```
```{r}
go_entrez <- go_entrez %>% mutate(coords = biomart_genes[match(.$symbol, biomart_genes$gene_name), "coords"],
                                  seqnames = biomart_genes[match(.$symbol, biomart_genes$gene_name), "seqnames"],
                                  start = biomart_genes[match(.$symbol, biomart_genes$gene_name), "start"],
                                  end = biomart_genes[match(.$symbol, biomart_genes$gene_name), "end"],
                                  n_cg_gene = biomart_genes[match(.$symbol, biomart_genes$gene_name), "num_cg"]) 
go_entrez %>% distinct(entrez, symbol, coords) %>% filter(is.na(coords)) %>% nrow(.)
```

* quick side note - how much does this impact??
```{r}
biomart_genes %>% filter(gene_name == "twnk")
biomart_genes %>% filter(ensembl_gene_id == "ENSG00000107815")
```
* same gene - different names
```{r}
go_entrez %>% filter(symbol == "C10orf2")
```

* this could be the case for all of them
 https://support.bioconductor.org/p/9148989/
* no clear solution
* as basically everyone hates this kind of stuff - read a recc which was don't switch just stick to what works for you
* maybe mapIds - different function
```{r}
go_entrez <- go_entrez %>% group_by(go) %>% mutate(n_genes_go = n()) %>% ungroup()
```

```{r}
go_entrez_filt <- go_entrez %>% filter(!is.na(start)) %>% group_by(go) %>% mutate(n_genes_go = n(), av_cg_go = mean(n_cg_gene), sum_cg_go = sum(n_cg_gene), gene_prop = sum(n_cg_gene)/n()) %>% ungroup()
```


```{r}
go_entrez_filt %>% distinct(go, term, n_genes_go, av_cg_go) %>%
  ggplot(aes(x = n_genes_go, y = av_cg_go)) +
  geom_point(alpha = 0.3)
```

```{r}
go_entrez_filt %>% distinct(go, term, n_genes_go, sum_cg_go) %>%
  ggplot(aes(x = n_genes_go, y = sum_cg_go)) +
  geom_point(alpha = 0.3)

go_entrez_filt %>% distinct(go, term, n_genes_go, gene_prop) %>%
  ggplot(aes(x = n_genes_go, y = gene_prop)) +
  geom_point(alpha = 0.3)
```
```{r}
go_entrez_filt_simp <- go_entrez_filt %>% distinct(go, term, n_genes_go, av_cg_go, sum_cg_go, gene_prop)
go_entrez_filt_simp %>% filter(av_cg_go == 0)
```

```{r}
cor <- cor.test(go_entrez_filt_simp$n_genes_go, go_entrez_filt_simp$sum_cg_go)
cor$statistic
```

```{r}
go_entrez_filt_simp
```
* want to make like my bias plot
* which would be grouping categories based on the number of cpgs
* oh except for that we need the results
* but I could still do a slightly more grouped thing?

```{r}
go_entrez_filt_simp %>% group_by(n_genes_go) %>% mutate(n_cat = n()) %>%
  ggplot(aes(x = log2(n_genes_go), y = n_cat)) +
  geom_point()
```

```{r}
go_entrez_filt_simp %>% group_by(n_genes_go) %>% mutate(n_cat = n()) %>%
  ggplot(aes(x = n_genes_go, y = n_cat)) +
  geom_point() +
  scale_x_continuous(limits = c(0,500))

go_entrez_filt_simp %>% group_by(n_genes_go) %>% mutate(n_cat = n()) %>%
  ggplot(aes(x = n_genes_go, y = log2(n_cat))) +
  geom_point() +
  scale_x_continuous(limits = c(0,500))
```
* more categories with not many genes
```{r}
go_entrez_filt_simp %>% 
  ggplot(aes(x = log2(n_genes_go), y = av_cg_go)) +
  geom_point(alpha = 0.3)

go_entrez_filt_simp %>% 
  ggplot(aes(x = n_genes_go, y = av_cg_go)) +
  geom_point(alpha = 0.3) +
  scale_x_continuous(limits = c(0,500))

go_entrez_filt_simp %>% 
  ggplot(aes(x = n_genes_go, y = av_cg_go)) +
  geom_point(alpha = 0.3) +
  scale_x_continuous(limits = c(0,250))

go_entrez_filt_simp %>% 
  ggplot(aes(x = n_genes_go, y = log2(av_cg_go))) +
  geom_point(alpha = 0.3) +
  scale_x_continuous(limits = c(0,500))

go_entrez_filt_simp %>% 
  ggplot(aes(x = n_genes_go, y = log2(av_cg_go))) +
  geom_point(alpha = 0.3) +
  scale_x_continuous(limits = c(0,250))
```
```{r}
go_entrez_filt_simp %>%
  ggplot(aes(x = sum_cg_go, y = av_cg_go)) +
  geom_point()
```
```{r}
go_entrez_filt_simp %>% .[order(.$av_cg_go, decreasing = TRUE),]
go_entrez_filt_simp %>% .[order(.$sum_cg_go, decreasing = TRUE),]
go_entrez_filt_simp %>% .[order(.$gene_prop, decreasing = TRUE),]
go_entrez_filt_simp %>% .[order(.$n_genes_go, decreasing = TRUE),]
```
* not suprised terms with most genes, and sum cg are the same - and very vague
* the top gene prop and av cg are also the same - and these are just those weird genes that have heaps of cpgs in them - as they also have only 1 or 2 genes in them

```{r}
go_entrez_filt_simp %>%
  ggplot(aes(x = log2(n_genes_go))) +
  geom_density()

go_entrez_filt_simp %>%
  ggplot(aes(x = av_cg_go)) +
  geom_density()

go_entrez_filt_simp %>%
  ggplot(aes(x = log2(av_cg_go))) +
  geom_density()

go_entrez_filt_simp %>%
  ggplot(aes(x = sum_cg_go)) +
  geom_density()

go_entrez_filt_simp %>%
  ggplot(aes(x = log2(sum_cg_go))) +
  geom_density()

go_entrez_filt_simp %>%
  ggplot(aes(x = gene_prop)) +
  geom_density()

go_entrez_filt_simp %>%
  ggplot(aes(x = log2(gene_prop))) +
  geom_density()
```


```{r}
go_entrez_filt %>% distinct(go, term, n_genes_go, av_cg_go) %>% mutate(min(n_genes_go), max(n_genes_go), mean(n_genes_go), sd(n_genes_go))

go_entrez_filt %>% distinct(go, term, n_genes_go, av_cg_go) %>% group_by(n_genes_go) %>% mutate(n_same=n(), av_same = mean(av_cg_go)) %>%
  ggplot(aes(x = n_same, y = av_same)) +
  geom_point(alpha = 0.3)
```




```{r}
go_entrez_filt %>% distinct(go, term, n_genes_go, av_cg_go) %>% filter(n_genes_go <= 10000) %>%
  ggplot(aes(x = log10(n_genes_go), y = log10(av_cg_go))) +
  geom_point(alpha = 0.3)
```

# bias in results
* show that when running hypergeometric we get significant sets that have lots of cpgs?

```{r}
go_entrez_filt_simp
```


```{r}
saveRDS(go_entrez_filt_simp, "../output/go_cg_summ.rds")
```

# i know what i want to do
* which is: are the categories that have lots of cpgs closely related?
* or any kind of connection?

* so i need some way to know the hierarchy or similarity of terms
* maybe have an itty bit of this in exploreCpGGST - as an exploratory: and if any systems have more CpGs in them already
* it would be nice to have a hierachy map with all of this to see how the gene sets are related
(I know I looked into a few things back in the day, see if I can find them)

* can't find what the packages were that I tried to do hierachy stuff with
oh yeah it was [gosemsim](https://bioconductor.org/packages/release/bioc/html/GOSemSim.html)
and also just googled
[go.db](https://bioconductor.org/packages/release/data/annotation/html/GO.db.html) - this one can show all the child terms of a term
[viseago](https://bioconductor.org/packages/release/bioc/html/ViSEAGO.html) - another semantic similarity one
[rrvgo](https://www.bioconductor.org/packages/devel/bioc/vignettes/rrvgo/inst/doc/rrvgo.html)
I think what I'm interested in this stuff is looking at the results overall and ways to do that and summarise those more elegantly than top 10
and just general understanding
I think this visualising is another file lol

```{r}
renv::install("bioc::rrvgo")
library(rrvgo)
go_analysis <- read.delim(system.file("extdata/example.txt", package="rrvgo"))
simMatrix <- calculateSimMatrix(go_analysis$ID,
                                orgdb="org.Hs.eg.db",
                                ont="BP",
                                method="Rel")
```
```{r}
scores <- setNames(-log10(go_analysis$qvalue), go_analysis$ID)
reducedTerms <- reduceSimMatrix(simMatrix,
                                scores,
                                threshold=0.7,
                                orgdb="org.Hs.eg.db")
```
* uses gosemsim in backend to do the sim
```{r}
scores %>% length(.)
go_analysis %>% nrow(.)
scores
```
```{r}
go_analysis$ID %in% filter(go_entrez, ontology == "BP")$go %>% summary()
```

```{r}
rownames(simMatrix) %in% go_analysis$ID
```

```{r}
scatterPlot(simMatrix, reducedTerms)
```

```{r}
treemapPlot(reducedTerms)
```
* in this one - space used by term proportional to score
* so I could have this instead as the number of CpGs.

```{r}
?calculateSimMatrix
```

# try on all
```{r}
go_entrez_filt_simp
```
```{r}
bsseq_dmrs
```

```{r}
simMatrix <- calculateSimMatrix(go_entrez_filt_simp$go,
                                orgdb="org.Hs.eg.db",
                                ont="BP",
                                method="Rel")
```
* unsuprisingly when you want to run this on everything - it takes a lot longer
* the vignette says something about speeding it up but i think thats about rerunning it
* so re-run this when I get home
* because it is going to do what I want to do

* the reduced terms fn gives individusl cat stats
that i could add sum cg to
and look at it
```{r}
go_entrez %>% distinct(ontology)
```
