==> analysis/00_longReadAnalysis.Rmd <==
---
title: "00_longReadAnalysis"
author: "Caitlin Page"
date: "2025-02-14"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: inline
---

## Introduction

```{r}
library(NanoMethViz)
library(plyranges)
library(dplyr)
library(ggplot2)
library(stringr)
library(bsseq)
library(dmrseq)
library(org.Mm.eg.db)
library(TxDb.Mmusculus.UCSC.mm10.knownGene)
```

```{r}
if (!dir.exists("../data/long_read/bams")) {
    options(timeout = 600)
    download.file("https://zenodo.org/records/12747551/files/input.tar.gz?download=1", "input.tar.gz")
    utils::untar("input.tar.gz")
    file.remove("input.tar.gz")
}
```
```{r}
bam_files <- dir("../data/long_read/input", pattern = "*bam$", full.names = TRUE)
bam_files
```

```{r}
samples <- read.table("../data/long_read/input/sample_anno.tsv", header = TRUE)
head(samples)
```
* they just used exon 7 - that only left very few significant DMRs
* workshop data is just chr7

```{r}
exon_anno <- get_exons_mm10() 
head(exon_anno)
```

```{r}
mbr <- ModBamResult(
    methy = ModBamFiles(
        paths = bam_files,
        samples = samples$sample
    ),
    samples = samples,
    exons = exon_anno,
    mod_code = "m"
)

plot_gene(mbr, "Peg3")
```


```{r}
nmr <- NanoMethResult(
    methy = "../data/long_read/methy.tsv.bgz",
    samples = samples,
    exons = exon_anno
)
str(nmr)
```

```{r}
plot_gene(nmr, "Peg3")
```

```{r}
bss <- methy_to_bsseq(nmr)

pat_cols <- str_detect(colnames(getCoverage(bss)), "pat")
mat_cols <- str_detect(colnames(getCoverage(bss)), "mat")

low_cov <- (rowSums(getCoverage(bss)[, pat_cols] == 0) == 3) |
           (rowSums(getCoverage(bss)[, mat_cols] == 0) == 3)

head(table(low_cov)/sum(table(low_cov)))
```

```{r}
bss <- bss[!low_cov, ]
```
```{r}
cgi_anno <- readRDS("../data/long_read/cgi_mm10.rds")
nrow(cgi_anno)
head(cgi_anno)
```


```{r}
lmr <- bsseq_to_log_methy_ratio(
    bss,
    regions = cgi_anno
)

groups <- colnames(lmr) %>%
    str_extract("mat|pat")

plot_mds(lmr, groups = groups)
```
* not strong differences, don't expect much in way of DM results

```{r}
gene_anno <- exons_to_genes(exons(nmr))
plot_agg_regions(nmr, regions = slice_sample(gene_anno, n = 100), group_col = "group")
```
```{r}
nrow(gene_anno)
head(gene_anno)
```

```{r}
plot_agg_regions(nmr, cgi_anno, group_col = "group")
```

```{r}
pData(bss)$condition <- NanoMethViz::samples(nmr)$group
```

```{r}
long_regions <- dmrseq(bss, testCovariate = "condition", minNumRegion = 20)
```

```{r}
nrow(long_regions)
head(long_regions)
```

```{r}
gene_anno_gr <- as_granges(dplyr::rename(gene_anno, seqnames = "chr"))

gene_anno_gr_tss <- gene_anno_gr %>%
    anchor_5p() %>%
    mutate(width = 1) %>%
    stretch(10000)

gene_dmr_overlaps <- join_overlap_intersect(long_regions, gene_anno_gr_tss)
nrow(gene_dmr_overlaps)
head(gene_dmr_overlaps)
```

```{r}
dmr_regions <- as_tibble(gene_dmr_overlaps) %>%
    dplyr::rename(chr = "seqnames")

dmr_regions <- dmr_regions %>% data.frame()

long_regions_sig <- dmr_regions %>%
    filter(qval < 0.05)

long_regions_sig %>% as.data.frame() 
```

* only 8 genes significant
* that's with TSS overlap
* let's do a full gene overlap

```{r}
gene_anno$seqnames <- gene_anno$chr
```

```{r}
long_gene_overlap <- find_overlaps(long_regions, as_granges(gene_anno), maxgap = 300) %>% data.frame()
long_gene_overlap %>% distinct(symbol)
```
* maxgap 300 to get a few extra genes
```{r}
long_gene_overlap %>%
    filter(qval < 0.05)
```
```{r}
gene_anno <- data.frame(gene_anno)
```

```{r}
long_genes_tested <- long_gene_overlap[,c(16,15,17)] %>% distinct(chr, gene_id, symbol) %>% 
  mutate(signif = ifelse(symbol %in% filter(long_gene_overlap, qval < 0.05)$symbol, 1, 0))
long_genes_tested$start <- gene_anno[match(long_genes_tested$symbol, gene_anno$symbol), "start"]
long_genes_tested$end <- gene_anno[match(long_genes_tested$symbol, gene_anno$symbol), "end"]
long_genes_tested$seqnames <- long_genes_tested$chr
long_genes_tested
```

```{r}
cgi_anno$seqnames <- cgi_anno$chr
```

```{r}
long_cg <- find_overlaps(as_granges(long_genes_tested), as_granges(cgi_anno), maxgap = 300) %>% data.frame() %>% 
  group_by(gene_id.x, symbol.x, start, end, signif) %>% summarise(n_cg = sum(cpgNum)) %>% ungroup() %>% data.frame()
```
```{r}
long_genes_tested$n_cg <- long_cg[match(long_genes_tested$symbol, long_cg$symbol.x), "n_cg"]
long_genes_tested <- long_genes_tested %>% mutate(n_cg = ifelse(is.na(n_cg), 0, n_cg))
long_genes_tested
```
* I have significant genes without any CG
* come back to this because something is wrong
```{r}
gene_anno %>% filter(chr == "chr7", symbol == "Peg3")
```


## my way of getting cg annotation
```{r get-cg}
source("../code/runBsseqDMR.R")
library(BSgenome.Mmusculus.UCSC.mm10)
seqnames(BSgenome.Mmusculus.UCSC.mm10)
cg_mouse <- get_cg_sites(BSgenome.Mmusculus.UCSC.mm10, 1, 19)
```

```{r}
long_cg <- find_overlaps(as_granges(long_genes_tested), as_granges(cg_mouse)) %>% data.frame() %>% 
  group_by(gene_id, symbol, n_cg) %>% summarise(num_cg = n()) %>% ungroup() %>% data.frame()

long_genes_tested$num_cg <- long_cg[match(long_genes_tested$symbol, long_cg$symbol), "num_cg"]
long_genes_tested <- long_genes_tested %>% mutate(num_cg = ifelse(is.na(num_cg), 0, num_cg))
long_genes_tested
```
* yeah this covers them better - replace it

* redoing with top 50 genes
```{r}
head(dmr_regions)
```
* should it be all the genes on chr7? 
* check how the data looks beforehand again
```{r}
dmr_regions$rank <- 1:nrow(dmr_regions)
```

```{r}
gene_anno_7 <- gene_anno %>% filter(seqnames == "chr7")
long_cg <- find_overlaps(as_granges(gene_anno_7), as_granges(cg_mouse)) %>% data.frame() %>% 
  group_by(gene_id, symbol) %>% summarise(num_cg = n()) %>% ungroup() %>% data.frame()

gene_anno_7$num_cg <- long_cg[match(gene_anno_7$symbol, long_cg$symbol), "num_cg"]
gene_anno_7 <- gene_anno_7 %>% mutate(num_cg = ifelse(is.na(num_cg), 0, num_cg), 
                                      qval = dmr_regions[match(.$symbol, dmr_regions$symbol), "qval"],
                                      rank = dmr_regions[match(.$symbol, dmr_regions$symbol), "rank"])
head(gene_anno_7)
```

## Gene set testing - GOSEQ

```{r}
gene.vector <- long_genes_tested$signif
names(gene.vector) <- long_genes_tested$gene_id

pwf <- goseq::nullp(gene.vector, "mm10", "knownGene", bias.data = long_genes_tested$num_cg)
GO.long <- goseq::goseq(pwf, "mm10", "knownGene")
GO.long <- GO.long %>% dplyr::mutate(FDR = stats::p.adjust(.data$over_represented_pvalue, method = "BH"))
head(GO.long)
```
* ok the FDR is 1 but the top term is genomic imprinting which is correct 


```{r}
gene_anno_7 <- gene_anno_7 %>% mutate(qval = ifelse(is.na(qval), 1.1, qval), signif = ifelse(qval < 0.05, 1, 0))
gene.vector <- gene_anno_7$signif
names(gene.vector) <- gene_anno_7$gene_id

pwf <- goseq::nullp(gene.vector, "mm10", "knownGene", bias.data = gene_anno_7$num_cg)
GO.long <- goseq::goseq(pwf, "mm10", "knownGene")
GO.long <- GO.long %>% dplyr::mutate(FDR = stats::p.adjust(.data$over_represented_pvalue, method = "BH"))
head(GO.long)
```

```{r}
gene_anno_7 <- gene_anno_7 %>% mutate(rank = ifelse(is.na(rank), nrow(.) + 1, rank), signif = ifelse(rank <= 50, 1, 0))
gene.vector <- gene_anno_7$signif
names(gene.vector) <- gene_anno_7$gene_id

pwf <- goseq::nullp(gene.vector, "mm10", "knownGene", bias.data = gene_anno_7$num_cg)
GO.long.50 <- goseq::goseq(pwf, "mm10", "knownGene")
GO.long.50 <- GO.long.50 %>% dplyr::mutate(FDR = stats::p.adjust(.data$over_represented_pvalue, method = "BH"))
head(GO.long.50)
```
* yeah ok taking top 50 genes made it worse


# null version
```{r}
gene.vector <- long_genes_tested$signif
names(gene.vector) <- long_genes_tested$gene_id

pwf <- goseq::nullp(gene.vector, "mm10", "knownGene", bias.data = long_genes_tested$num_cg)

GO.long_null <- goseq::goseq(pwf, "mm10", "knownGene", method = "Hypergeometric")
GO.long_null <- GO.long_null %>% dplyr::mutate(FDR = stats::p.adjust(.data$over_represented_pvalue, method = "BH"))
head(GO.long_null)
```
* also imprinting

```{r}
gene_anno_7 <- gene_anno_7 %>% mutate(qval = ifelse(is.na(qval), 1.1, qval), signif = ifelse(qval < 0.05, 1, 0))
gene.vector <- gene_anno_7$signif
names(gene.vector) <- gene_anno_7$gene_id

pwf <- goseq::nullp(gene.vector, "mm10", "knownGene", bias.data = gene_anno_7$num_cg)
GO.long_null <- goseq::goseq(pwf, "mm10", "knownGene", method = "Hypergeometric")
GO.long_null <- GO.long_null %>% dplyr::mutate(FDR = stats::p.adjust(.data$over_represented_pvalue, method = "BH"))
head(GO.long_null)
```


```{r}
gene_anno_7 <- gene_anno_7 %>% mutate(rank = ifelse(is.na(rank), nrow(.) + 1, rank), signif = ifelse(rank <= 50, 1, 0))
gene.vector <- gene_anno_7$signif
names(gene.vector) <- gene_anno_7$gene_id

pwf <- goseq::nullp(gene.vector, "mm10", "knownGene", bias.data = gene_anno_7$num_cg)
GO.long.50_null <- goseq::goseq(pwf, "mm10", "knownGene", method = "Hypergeometric")
GO.long.50_null <- GO.long.50_null %>% dplyr::mutate(FDR = stats::p.adjust(.data$over_represented_pvalue, method = "BH"))
head(GO.long.50_null)
```

## fisher

```{r}
gene.vector <- long_genes_tested$signif
names(gene.vector) <- long_genes_tested$gene_id

pwf <- goseq::nullp(gene.vector, "mm10", "knownGene", bias.data = long_genes_tested$num_cg)
pwf
GO.long_fish <- goseq_mod(pwf, "mm10", "knownGene", method = "Fishers")
GO.long_fish
GO.long_fish <- GO.long_fish %>% dplyr::mutate(FDR = stats::p.adjust(.data$over_represented_pvalue, method = "BH"))
GO.long_fish
```

```{r}
gene_anno_7 <- gene_anno_7 %>% mutate(rank = ifelse(is.na(rank), nrow(.) + 1, rank), signif = ifelse(rank <= 50, 1, 0))
gene.vector <- gene_anno_7$signif
names(gene.vector) <- gene_anno_7$gene_id

pwf <- goseq::nullp(gene.vector, "mm10", "knownGene", bias.data = gene_anno_7$num_cg)
pwf
GO.long.50_fish <- goseq_mod(pwf, "mm10", "knownGene", method = "Fishers")
GO.long.50_fish
GO.long.50_fish <- GO.long.50_fish %>% dplyr::mutate(FDR = stats::p.adjust(.data$over_represented_pvalue, method = "BH"))
GO.long.50_fish
```


## now for my function

```{r}
ggVennDiagram(list(wall = filter(GO.long, over_represented_pvalue < 0.05)$category, 
                   null = filter(GO.long_null, over_represented_pvalue < 0.05)$category))
```
