==> analysis/01_exploreBiasByBins.Rmd <==
---
title: "01_exploreBiasByBins"
author: "Caitlin Page"
date: "2024-09-05"
output: workflowr::wflow_html
editor_options: 
  chunk_output_type: inline
---
chunk_output_type: console
## Introduction

```{r}
library(plyranges)
library(dplyr)
library(ggplot2)
library(reshape2)
library(ChIPseqSpikeInFree)
library(biovizBase)
library(BSgenome.Hsapiens.UCSC.hg19)
```

```{r}
wgbs_counts <- readRDS("../output/wgbs_counts.rds")
bsseq_res <- readRDS("../output/bsseq_res.rds")
bsseq_dmrs <- readRDS("../output/bsseq_dmrs.rds")
dmrcate_seq_dmr <- readRDS("../output/dmrcate_seq_dmr.rds") %>% data.frame()
biomart_genes <- readRDS("../output/genes_biomart.rds")
```
```{r}
bsseq_dmrs_filt <- bsseq_dmrs %>% filter(abs(areaStat) > 100) # filter to significant DMRs
```

# What bin size to choose?

```{r}
bsseq_dmrs %>%
  ggplot(aes(x = width, colour = seqnames)) +
  geom_density()

bsseq_dmrs_filt %>%
  ggplot(aes(x = width, colour = seqnames)) +
  geom_density()
```

* dmrs are mostly < 2000 bp

```{r shared-code}
source("../code/function_binBiasPlot.R")
plotBiasGenomeBins(wgbs_counts, bsseq_dmrs, bin_size = 2000)

plotBiasGenomeBins(wgbs_counts, bsseq_dmrs_filt, bin_size = 2000)
```

* explore other bin sizes
```{r}
source("../code/function_binBiasPlot.R")
plotBiasGenomeBins(wgbs_counts, dmrcate_seq_dmr, bin_size = 2000)
```


```{r}
source("../code/function_binBiasPlot.R")
plotBiasGenomeBins(wgbs_counts, bsseq_dmrs, bin_size = 500)
plotBiasGenomeBins(wgbs_counts, bsseq_dmrs_filt, bin_size = 500)
```

```{r}
source("../code/function_binBiasPlot.R")
plotBiasGenomeBins(wgbs_counts, bsseq_dmrs, bin_size = 1000)
plotBiasGenomeBins(wgbs_counts, bsseq_dmrs_filt, bin_size = 1000)
```

```{r}
source("../code/function_binBiasPlot.R")
plotBiasGenomeBins(wgbs_counts, bsseq_dmrs, bin_size = 5000)
plotBiasGenomeBins(wgbs_counts, bsseq_dmrs_filt, bin_size = 5000)
```

* prefer 2000 bp bin length

```{r}
source("../code/function_binBiasPlot.R")
genome_bins <- plotBiasGenomeBins(wgbs_counts, bsseq_dmrs, bin_size = 2000) 
genome_bins <- genome_bins[[1]]
```

Questions:
* why are some bins with high cgs have no dmrs?

# Strange proportions in big bins
Possible explanations
* telomeres - strange things at the starts and the ends of chromosomes
Test: plot the first and last 20 bins per chromosome - are these the bins with high cg numbers?

* tail
```{r}
genome_bins %>% group_by(seqnames) %>% do(tail(., n = 20)) %>% mutate(n = 1:n()) %>%
  ggplot(aes(x = n, y = num_cg, colour = seqnames)) +
  geom_point()
```
* head
```{r}
genome_bins %>% group_by(seqnames) %>% do(head(., n = 20)) %>% mutate(n = 1:n()) %>%
  ggplot(aes(x = n, y = num_cg, colour = seqnames)) +
  geom_point()
```

Result: ends of chromosomes are not the source of high cg bins

## GC content?

```{r}
genome_bins <- cbind(genome_bins, biovizBase::GCcontent(BSgenome.Hsapiens.UCSC.hg19, as_granges(genome_bins))) %>% data.frame()
```

```{r}
genome_bins %>%
  ggplot(aes(x = num_cg, y = C.G, colour = has_dmr)) +
  geom_point(alpha = 0.3)
```

```{r}
genome_bins %>%
  ggplot(aes(x = C.G)) +
  geom_density()

genome_bins %>%
  ggplot(aes(x = C.G, colour = has_dmr)) +
  geom_density()
```

```{r}
cor.test(genome_bins$num_cg, genome_bins$C.G)
```

GC content is related to coverage

## Coverage in bins
```{r}
counts_per_bin <- find_overlaps(as_granges(genome_bins), as_granges(wgbs_counts)) %>% data.frame()
```

```{r}
counts_per_bin_sum <- counts_per_bin %>% group_by(num) %>% mutate(sum_nk_TM = sum(nk_C_TM), sum_nk_U1 = sum(nk_C_U1), sum_nk_UF = sum(nk_C_UF),
                                            sum_b_TX = sum(b_C_TX), sum_b_UB = sum(b_C_UB), sum_b_UR = sum(b_C_UR)) %>% distinct(bin_pos, num, sum_nk_TM, sum_nk_U1, sum_nk_UF, sum_b_TX, sum_b_UB, sum_b_UR)
counts_per_bin_sum$av_nk <- rowMeans(counts_per_bin_sum[,3:5])
counts_per_bin_sum$av_b <- rowMeans(counts_per_bin_sum[,6:8])
counts_per_bin_sum <- counts_per_bin_sum %>% ungroup() %>% data.frame()
```

```{r}
genome_bins <- genome_bins %>% mutate(av_nk = counts_per_bin_sum[match(.$bin_pos, counts_per_bin_sum$bin_pos), "av_nk"],
                                      av_b = counts_per_bin_sum[match(.$bin_pos, counts_per_bin_sum$bin_pos), "av_b"])
genome_bins[is.na(genome_bins)] <- 0
genome_bins
```

```{r}
genome_bins %>%
  ggplot(aes(y = av_nk, x = C.G)) +
  geom_point(colour = "red", alpha = 0.3) +
  geom_point(aes(y = av_b, x = C.G), colour = "black", alpha = 0.3)
```
```{r}
genome_bins %>%
  ggplot(aes(y = log2(av_nk), x = C.G)) +
  geom_point(colour = "red", alpha = 0.3) +
  geom_point(aes(y = log2(av_b), x = C.G), colour = "black", alpha = 0.3)
```

```{r}
genome_bins %>% mutate(num_cg_gr_200 = ifelse(num_cg > 200, TRUE, FALSE)) %>%
  ggplot(aes(y = av_nk, colour = num_cg_gr_200)) +
  geom_boxplot()

genome_bins %>% mutate(num_cg_gr_200 = ifelse(num_cg > 200, TRUE, FALSE)) %>%
  ggplot(aes(x = av_nk, colour = num_cg_gr_200)) +
  geom_density()
```

```{r}
genome_bins[1:5000,] %>%
  ggplot(aes(x = av_nk, y = av_b)) +
  geom_point()
```
```{r}
cor.test(genome_bins$av_nk, genome_bins$av_b)
```

## how common are these bins anyway?
```{r}
genome_bins %>%
  ggplot(aes(x = num_cg)) +
  geom_density()
genome_bins %>%
  ggplot(aes(x = num_cg, colour = has_dmr)) +
  geom_density()
```
############

```{r}
dmrcate_seq_dmr <- readRDS("../output/dmrcate_seq_dmr.rds") %>% data.frame()
```
```{r}
biomart_genes <- readRDS("../output/genes_biomart.rds")
biomart_genes <- biomart_genes %>% data.frame()
```

```{r}
source("../code/function_binBiasPlot.R")
test_anno <- annoGeneDmr(wgbs_counts, dmrcate_seq_dmr, "biomart", biomart_genes)
test_anno_filt <- annoGeneDmr(wgbs_counts, dmrcate_seq_dmr[1:1000,], "biomart", biomart_genes)

overlap_gene_dmr <- find_overlaps(as_granges(dmrcate_seq_dmr), as_granges(eh_genes)) %>% data.frame()
```


```{r}
source("../code/function_binBiasPlot.R")
plotBiasGeneLength(biomart_genes, test_anno_filt, bin_size = 100)
plotBiasGeneLength(biomart_genes, test_anno_filt, bin_size = 200)

plotBiasGeneLength(biomart_genes, test_anno, bin_size = 100)
plotBiasGeneLength(biomart_genes, test_anno, bin_size = 200)
```

```{r}
plotBiasGeneLength(biomart_genes, test_anno_filt, bin_size = 300)
```
```{r}
plotBiasGenomeBins()
```

```{r}
source("../code/function_binBiasPlot.R")
plotBiasGenomeBins(wgbs_counts, dmrcate_seq_dmr, bin_size = 5000, )
plotBiasGenomeBins(wgbs_counts, dmrcate_seq_dmr[1:1000,], bin_size = 5000)
```
```{r}
plotBiasGenomeBins(wgbs_counts, dmrcate_seq_dmr, bin_size = 500)
```
```{r}
dmrcate_seq_dmr
```

```{r}
test_anno
```

```{r}
test_anno2 <- annoGeneDmr(wgbs_counts, mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr)), "biomart", biomart_genes)
```

```{r}
find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(mutate(biomart_genes, gene_width = end - start + 1))) %>% data.frame() %>%
  ggplot(aes(x = gene_width, y = no.cpgs)) +
  geom_point()
```
```{r}
biomart_genes %>% mutate(gene_width = end - start + 1) %>% distinct(gene_name, gene_width, num_cg) %>%
  ggplot(aes(x = gene_width, y = num_cg)) +
  geom_point(alpha = 0.4) +
  geom_abline(slope = 1, intercept = 0, colour = "red")
```
```{r}
find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(mutate(biomart_genes, gene_width = end - start + 1))) %>% data.frame() %>% distinct(gene_name, gene_width, num_cg) %>%
  ggplot(aes(x = gene_width, y = num_cg)) +
  geom_point(alpha = 0.4) +
  geom_abline(slope = 1, intercept = 0, colour = "red")

find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(mutate(biomart_genes, gene_width = end - start + 1))) %>% data.frame() %>% distinct(gene_name, gene_width, num_cg) %>%
  ggplot(aes(x = log2(gene_width), y = log2(num_cg + 0.1))) +
  geom_point(alpha = 0.4) +
  geom_abline(slope = 1, intercept = 0, colour = "red")
```
```{r}
dmrcate_seq_dmr %>%
  ggplot(aes(x = width, y = no.cpgs)) +
  geom_point() +
  geom_abline(slope = 1, intercept = 0, colour = "red")
```
```{r}
cor.test(dmrcate_seq_dmr$width, dmrcate_seq_dmr$no.cpgs)
```
```{r}
gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(mutate(biomart_genes, gene_width = end - start + 1))) %>% data.frame() %>% distinct(gene_name, gene_width, num_cg)
cor.test(gene_try$gene_width, gene_try$num_cg)
```
* maybe I had only done the math not the visual
```{r}
biomart_genes %>% mutate(gene_width = end - start + 1) %>% distinct(gene_name, gene_width, num_cg) %>%
  ggplot(aes(x = log2(gene_width), y = log2(num_cg))) +
  geom_point(alpha = 0.4) +
  geom_abline(slope = 1, intercept = 0, colour = "red") 
```
log2 doesn't make it look better

```{r}
source("../code/mmGST.R")
source("../code/function_binBiasPlot.R")
run_miss_methyl_b(run_miss_methyl_a(biomart_genes, annoGeneDmr(wgbs_counts, dmrcate_seq_dmr[1:1000,], "biomart", biomart_genes)), "GO", "Wallenius")[[1]]
```


########

```{r}
.estimatePWF <- function(D,bias)
  # An alternative to goseq function nullp, which is transformation invariant
  # Belinda Phipson and Gordon Smyth
  # 6 March 2015
{
  prior.prob <- bias
  o <- order(bias)
  prior.prob[o] <- limma::tricubeMovingAverage(D[o],span=0.5)
  prior.prob
}

.plotBias <- function(D,bias)
  # Plotting function to show gene level CpG density bias
  # Belinda Phipson
  # 5 March 2015
{
  o <- order(bias)
  splitf <- rep(1:100,each=200)[1:length(bias)]
  avgbias <- tapply(bias[o],factor(splitf),mean)
  sumDM <- tapply(D[o],factor(splitf),sum)
  propDM <- sumDM/table(splitf)
  graphics::par(mar=c(5,5,2,2))
  graphics::plot(avgbias,as.vector(propDM),
                 xlab="Number of CpGs per gene (binned)",
                 ylab="Proportion Differential Methylation",cex.lab=1.5,
                 cex.axis=1.2)
  graphics::lines(stats::lowess(avgbias,propDM),col=4,lwd=2)
}
```
```{r}
freq_genes <- out$freq # num cpgs per gene
test.de <- out$de #gene de or not
```
```{r}
gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(mutate(biomart_genes, gene_width = end - start + 1))) %>% data.frame()
gene_info <- biomart_genes %>% mutate(width = end - start + 1, is_de = ifelse(gene_name %in% gene_try$gene_name, 1, 0)) %>% distinct(gene_name, num_cg, is_de, width)
```

```{r}
#test_pwf <- .estimatePWF(D=test.de,bias=as.vector(freq_genes))
.plotBias(D = gene_info$is_de, bias = as.vector(gene_info$num_cg))
.plotBias(D = gene_info$is_de, bias = as.vector(gene_info$width)) # width
```
```{r}
#library(GenomicFeatures)
txdb_genes <- genes(TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb_genes <- find_overlaps(txdb_genes, as_granges(wgbs_counts[,1:4])) %>% data.frame() %>% group_by(gene_id, seqnames, start, end, width, strand) %>% summarise(num_cg = n()) %>% ungroup() %>% data.frame()
```

```{r}
gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(txdb_genes)) %>% data.frame()
gene_info <- txdb_genes %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)
```

```{r}
.plotBias(D = gene_info$is_de, bias = as.vector(gene_info$num_cg))
.plotBias <- function(D,bias)
  # Plotting function to show gene level CpG density bias
  # Belinda Phipson
  # 5 March 2015
{
  o <- order(bias)
  splitf <- rep(1:100,each=200)[1:length(bias)]
  avgbias <- tapply(bias[o],factor(splitf),mean)
  sumDM <- tapply(D[o],factor(splitf),sum)
  propDM <- sumDM/table(splitf)
  graphics::par(mar=c(5,5,2,2))
  graphics::plot(avgbias,as.vector(propDM),
                 xlab="Length of gene (binned)",
                 ylab="Proportion Differential Methylation",cex.lab=1.5,
                 cex.axis=1.2)
  graphics::lines(stats::lowess(avgbias,propDM),col=4,lwd=2)
}
.plotBias(D = gene_info$is_de, bias = as.vector(gene_info$width))
```
* so so much nicer

```{r}
txdb_genes %>%
  ggplot(aes(x = width, y = num_cg)) +
  geom_point(alpha = 0.4) +
  geom_abline(slope = 1, intercept = 0, colour = "red")

txdb_genes %>%
  ggplot(aes(x = width, y = num_cg)) +
  geom_point(alpha = 0.4) +
  geom_abline(slope = 1, intercept = 0, colour = "red") +
  scale_x_continuous(limits = c(0, 2500000)) +
  scale_y_continuous(limits = c(0,30000))

txdb_genes %>%
  ggplot(aes(x = width, y = num_cg)) +
  geom_point(alpha = 0.4) +
  geom_abline(slope = 1, intercept = 0, colour = "red") +
  scale_x_continuous(limits = c(0, 1000000)) +
  scale_y_continuous(limits = c(0,20000))
```
```{r}
# just to limits in the 2 bias plots
txdb_genes %>%
  ggplot(aes(x = width, y = num_cg)) +
  geom_point(alpha = 0.4) +
  geom_abline(slope = 1, intercept = 0, colour = "red") +
  scale_x_continuous(limits = c(0, 150000)) +
 scale_y_continuous(limits = c(0,6000))
```
```{r}
txdb_genes %>%
  ggplot(aes(x = width, y = num_cg)) +
  geom_point(alpha = 0.4) +
  geom_abline(slope = 1, intercept = 0, colour = "red") +
  scale_x_continuous(limits = c(0, 120000)) +
 scale_y_continuous(limits = c(0,4500)) +
  labs(x = "Length of gene",
       y = "CpGs overlapping gene") +
  theme_bw()
```

```{r}
txdb_genes[,c(1,5,7)] %>% melt(id.vars = "gene_id") %>%
  ggplot(aes(x = log2(value), fill = variable)) +
  geom_density(alpha = 0.4)
```
```{r}
2^8
```
```{r}
txdb_genes %>% filter(width > 32, width < 256)
```


```{r}
txdb_genes %>%
  ggplot(aes(x = log2(width))) +
  geom_density()

txdb_genes %>%
  ggplot(aes(x = log2(num_cg))) +
  geom_density()
```


#top 1000

```{r}
gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr[1:1000,], id = 1:1000)), as_granges(txdb_genes)) %>% data.frame()
gene_info <- txdb_genes %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)
.plotBias <- function(D,bias)
  # Plotting function to show gene level CpG density bias
  # Belinda Phipson
  # 5 March 2015
{
  o <- order(bias)
  splitf <- rep(1:100,each=200)[1:length(bias)]
  avgbias <- tapply(bias[o],factor(splitf),mean)
  sumDM <- tapply(D[o],factor(splitf),sum)
  propDM <- sumDM/table(splitf)
  graphics::par(mar=c(5,5,2,2))
  graphics::plot(avgbias,as.vector(propDM),
                 xlab="Number of CpGs per gene (binned)",
                 ylab="Proportion Differential Methylation",cex.lab=1.5,
                 cex.axis=1.2)
  graphics::lines(stats::lowess(avgbias,propDM),col=4,lwd=2)
}
.plotBias(D = gene_info$is_de, bias = as.vector(gene_info$num_cg))
.plotBias <- function(D,bias)
  # Plotting function to show gene level CpG density bias
  # Belinda Phipson
  # 5 March 2015
{
  o <- order(bias)
  splitf <- rep(1:100,each=200)[1:length(bias)]
  avgbias <- tapply(bias[o],factor(splitf),mean)
  sumDM <- tapply(D[o],factor(splitf),sum)
  propDM <- sumDM/table(splitf)
  graphics::par(mar=c(5,5,2,2))
  graphics::plot(avgbias,as.vector(propDM),
                 xlab="Length of gene (binned)",
                 ylab="Proportion Differential Methylation",cex.lab=1.5,
                 cex.axis=1.2)
  graphics::lines(stats::lowess(avgbias,propDM),col=4,lwd=2)
}
.plotBias(D = gene_info$is_de, bias = as.vector(gene_info$width))
```
* better with all


```{r}
#library(GenomicFeatures)
txdb_genes <- genes(TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb_genes <- find_overlaps(txdb_genes, as_granges(wgbs_counts[,1:4])) %>% data.frame() %>% group_by(gene_id, seqnames, start, end, width, strand) %>% summarise(num_cg = n()) %>% ungroup() %>% data.frame()
```

```{r}
gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(txdb_genes)) %>% data.frame()
gene_info <- txdb_genes %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)
```

```{r}
run_test <- list(
  sig.eg = gene_info %>% filter(is_de == 1) %>% .$gene_id,
  universe = gene_info$gene_id,
  freq = gene_info$num_cg,
  de = gene_info$is_de,
  equiv = gene_info$num_cg
)
```

```{r}
gst <- run_miss_methyl_b(run_test, method = "Wallenius")
gst_fisher <- run_miss_methyl_b(run_test, method = "Fishers")
gst_hyper <- run_miss_methyl_b(run_test, method = "None")
```
```{r}
library(UpSetR)
```

```{r}
upset_input <- list(GOMethSeq = gst[[1]] %>% filter(FDR < 0.05) %>% .$GOID,
                    Fishers = gst_fisher[[1]] %>% filter(FDR < 0.05) %>% .$GOID,
                    Biased = gst_hyper[[1]] %>% filter(FDR < 0.05) %>% .$GOID)
upset(fromList(upset_input), order.by = "freq", text.scale = 2)
```

```{r}
gst <-gst[[1]] %>% mutate(rank_fdr = 1:n())
gst_fisher <- gst_fisher[[1]] %>% mutate(rank_fdr = 1:n())
gst_hyper <- gst_hyper[[1]] %>% mutate(rank_fdr = 1:n())
```

```{r}
go_entrez <- readRDS("../output/go_entrez.rds")
go_entrez <- go_entrez[,1:2]
gene_info
go_entrez <- go_entrez %>%
  mutate(num_cg = gene_info[match(.$entrez, gene_info$num_cg), "num_cg"],
         gomethseq = ifelse(go %in% filter(gst, FDR < 0.05)$GOID, TRUE, FALSE),
         fishers = ifelse(go %in% filter(gst_fisher, FDR < 0.05)$GOID, TRUE, FALSE),
         biased = ifelse(go %in% filter(gst_hyper, FDR < 0.05)$GOID, TRUE, FALSE),
         rank_gomethseq = gst[match(.$go, gst$GOID), "rank_fdr"],
         rank_fishers = gst_fisher[match(.$go, gst_fisher$GOID), "rank_fdr"],
         rank_biased = gst_hyper[match(.$go, gst_hyper$GOID), "rank_fdr"])
go_entrez <- go_entrez %>%
  group_by(go) %>%
  mutate(mean_num_cg = mean(num_cg, na.rm = TRUE)) %>%
  ungroup() %>% data.frame()
```


```{r}
go_entrez %>% distinct(go, mean_num_cg, rank_gomethseq, rank_biased) %>%
  filter(rank_gomethseq <= 1000, rank_biased <= 1000) %>%
  ggplot(aes(x = mean_num_cg, y = rank_biased - rank_gomethseq)) + 
  geom_point() +
  geom_hline(yintercept = 0, colour = "red", linetype = "dashed") +
  labs(x = "Average No. CpGs in GO category", y = "Increase in rank in Wallenius") 
 # scale_x_continuous(limits = c(0, 1500))
```

```{r}
# find the unique categories
go_entrez %>% distinct(go, mean_num_cg, gomethseq, biased) %>%
  filter(gomethseq != biased) %>% 
  mutate(Method = ifelse(gomethseq == FALSE, "Biased", "GOMethSeq")) %>%
  ggplot(aes(x = reorder(Method, mean_num_cg), y = mean_num_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(x = "Method", y = "Average number of CpGs per ontology category")

# old - all
go_entrez %>% distinct(go, mean_num_cg, gomethseq, biased) %>%
  filter(gomethseq == TRUE | biased == TRUE) %>% 
  mutate(Method = case_when(gomethseq == FALSE & biased == TRUE ~ "Biased",
                            gomethseq == TRUE & biased == TRUE ~ "Both",
                            TRUE ~ "Bias Corrected")) %>%
  ggplot(aes(x = reorder(Method, mean_num_cg), y = mean_num_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(x = "Method", y = "Average number of CpGs per ontology category")
```
```{r}
immuneGO <- read.table("../data/gene_sets/GO-immune-system-process.txt", header = FALSE)
names(immuneGO) <- "GOID"
immuneGO <- distinct(immuneGO)
```


```{r}
truth_gst_100 <- rbind(
  gst[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")])

truth_gst_100 %>%
  ggplot(aes(x = rank_fdr, y = in_imm, colour = method)) +
  geom_line(size = 2) +
  labs(x = "Rank", y = "Cumulative no. in immune sets") +
  theme_bw()
```

* top 1k
```{r}
gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr[1:1000,], id = 1:1000)), as_granges(txdb_genes)) %>% data.frame()
gene_info <- txdb_genes %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)
```

```{r}
run_test <- list(
  sig.eg = gene_info %>% filter(is_de == 1) %>% .$gene_id,
  universe = gene_info$gene_id,
  freq = gene_info$num_cg,
  de = gene_info$is_de,
  equiv = gene_info$num_cg
)
```

```{r}
gst_1k <- run_miss_methyl_b(run_test, method = "Wallenius")
gst_1k_fisher <- run_miss_methyl_b(run_test, method = "Fishers")
gst_1k_hyper <- run_miss_methyl_b(run_test, method = "None")
```


```{r}
upset_input <- list(GOMethSeq = gst_1k[[1]] %>% filter(FDR < 0.05) %>% .$GOID,
                    Fishers = gst_1k_fisher[[1]] %>% filter(FDR < 0.05) %>% .$GOID,
                    Biased = gst_1k_hyper[[1]] %>% filter(FDR < 0.05) %>% .$GOID)
upset(fromList(upset_input), order.by = "freq", text.scale = 2)
```
```{r}
gst_1k <-gst_1k[[1]] %>% mutate(rank_fdr = 1:n())
gst_1k_fisher <- gst_1k_fisher[[1]] %>% mutate(rank_fdr = 1:n())
gst_1k_hyper <- gst_1k_hyper[[1]] %>% mutate(rank_fdr = 1:n())
```

```{r}
go_entrez <- readRDS("../output/go_entrez.rds")
go_entrez <- go_entrez[,1:2]
gene_info
go_entrez <- go_entrez %>%
  mutate(num_cg = gene_info[match(.$entrez, gene_info$num_cg), "num_cg"],
         gomethseq = ifelse(go %in% filter(gst_1k, FDR < 0.05)$GOID, TRUE, FALSE),
         fishers = ifelse(go %in% filter(gst_1k_fisher, FDR < 0.05)$GOID, TRUE, FALSE),
         biased = ifelse(go %in% filter(gst_1k_hyper, FDR < 0.05)$GOID, TRUE, FALSE),
         rank_gomethseq = gst_1k[match(.$go, gst_1k$GOID), "rank_fdr"],
         rank_fishers = gst_1k_fisher[match(.$go, gst_1k_fisher$GOID), "rank_fdr"],
         rank_biased = gst_1k_hyper[match(.$go, gst_1k_hyper$GOID), "rank_fdr"])
go_entrez <- go_entrez %>%
  group_by(go) %>%
  mutate(mean_num_cg = mean(num_cg, na.rm = TRUE)) %>%
  ungroup() %>% data.frame()
         
```


```{r}
go_entrez %>% distinct(go, mean_num_cg, rank_gomethseq, rank_biased) %>%
  filter(rank_gomethseq <= 1000, rank_biased <= 1000) %>%
  ggplot(aes(x = mean_num_cg, y = rank_biased - rank_gomethseq)) + 
  geom_point() +
  geom_hline(yintercept = 0, colour = "red", linetype = "dashed") +
  labs(x = "Average No. CpGs in GO category", y = "Increase in rank in Wallenius") 
 # scale_x_continuous(limits = c(0, 1500))
```


```{r}
go_entrez %>% distinct(go, mean_num_cg, gomethseq, biased) %>%
  filter(gomethseq != biased) %>% 
  mutate(Method = ifelse(gomethseq == FALSE, "Biased", "GOMethSeq")) %>%
  ggplot(aes(x = reorder(Method, mean_num_cg), y = mean_num_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(x = "Method", y = "Average number of CpGs per ontology category")

go_entrez %>% distinct(go, mean_num_cg, gomethseq, biased) %>%
  filter(gomethseq == TRUE | biased == TRUE) %>% 
  mutate(Method = case_when(gomethseq == FALSE & biased == TRUE ~ "Biased",
                            gomethseq == TRUE & biased == TRUE ~ "Both",
                            TRUE ~ "Bias Corrected")) %>%
  ggplot(aes(x = reorder(Method, mean_num_cg), y = mean_num_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(x = "Method", y = "Average number of CpGs per ontology category")
```


```{r}
truth_gst_100 <- rbind(
  gst[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")])

truth_gst_100 %>%
  ggplot(aes(x = rank_fdr, y = in_imm, colour = method)) +
  geom_line(size = 2) +
  labs(x = "Rank", y = "Cumulative no. in immune sets") +
  theme_bw()
```
```{r}
txdb_genes %>% summarise(sum(num_cg))
```

* t cell
```{r}
gst_t <- gst_t[[1]]
```

```{r}
truth_gst_100 <- rbind(
  gst_t[1:100,] %>% mutate(rank_fdr = 1:100, in_imm = cumsum(GOID %in% immuneGO$GOID), method = "GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_t_hyper[1:100,] %>% mutate(rank_fdr = 1:100, in_imm = cumsum(GOID %in% immuneGO$GOID), method = "Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")])

truth_gst_100 %>%
  ggplot(aes(x = rank_fdr, y = in_imm, colour = method)) +
  geom_line(size = 2) +
  labs(x = "Rank", y = "Cumulative no. in immune sets") +
  theme_bw()
```


## more bias plots

```{r}
.plotBias <- function(D,bias)
  # Plotting function to show gene level CpG density bias
  # Belinda Phipson
  # 5 March 2015
{
  o <- order(bias)
  splitf <- rep(1:100,each=200)[1:length(bias)]
  avgbias <- tapply(bias[o],factor(splitf),mean)
  sumDM <- tapply(D[o],factor(splitf),sum)
  propDM <- sumDM/table(splitf)
  graphics::par(mar=c(5,5,2,2))
  graphics::plot(avgbias,as.vector(propDM),
                 xlab="Number of CpGs per gene (binned)",
                 ylab="Proportion Differential Methylation",cex.lab=1.5,
                 cex.axis=1.2)
  graphics::lines(stats::lowess(avgbias,propDM),col=4,lwd=2)
}
```


```{r}
txdb_genes_mod <- find_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         start = ifelse(strand == "+", start - 1500, start),
         end = ifelse(strand == "+", end, end + 1500)) %>% .[,c(1:4,6)] %>% as_granges(), 
  as_granges(wgbs_counts[,1:4])) %>% data.frame() %>%
  group_by(gene_id, seqnames, start, end, strand, width) %>%
  summarise(num_cg = n()) %>% ungroup() %>% data.frame()

gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(txdb_genes_mod)) %>% data.frame()
gene_info <- txdb_genes_mod %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)

.plotBias(D = gene_info$is_de, bias = as.vector(gene_info$num_cg))
```

```{r}
txdb_genes_mod <- find_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         start = ifelse(strand == "+", start - 2000, end - 2000),
         end = ifelse(strand == "+", start + 4000, end + 2000)) %>% .[,c(1:4,6)] %>% as_granges(), 
  as_granges(wgbs_counts[,1:4])) %>% data.frame() %>%
  group_by(gene_id, seqnames, start, end, strand, width) %>%
  summarise(num_cg = n()) %>% ungroup() %>% data.frame()

gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(txdb_genes_mod)) %>% data.frame()
gene_info <- txdb_genes_mod %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)

.plotBias(D = gene_info$is_de, bias = as.vector(gene_info$num_cg))
```

```{r}
txdb_genes_mod <- find_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         start = ifelse(strand == "+", start - 2000, end - 2000),
         end = ifelse(strand == "+", start + 4000, end + 2000)) %>% .[,c(1:4,6)] %>% as_granges(), 
  as_granges(wgbs_counts[,1:4])) %>% data.frame() %>%
  group_by(gene_id, seqnames, start, end, strand, width) %>%
  summarise(num_cg = n()) %>% ungroup() %>% data.frame()
gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(txdb_genes_mod)) %>% data.frame()
gene_info <- txdb_genes_mod %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)
gene_info %>% group_by(is_de) %>% summarise(n=n())
```
```{r}
txdb_genes_mod <- find_overlaps(txdb_genes %>% as_granges(), 
  as_granges(wgbs_counts[,1:4])) %>% data.frame() %>%
  group_by(gene_id, seqnames, start, end, strand, width) %>%
  summarise(num_cg = n()) %>% ungroup() %>% data.frame()
gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(txdb_genes_mod)) %>% data.frame()
gene_info <- txdb_genes_mod %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)
gene_info %>% group_by(is_de) %>% summarise(n=n())
```
```{r}
1305/4706
```


```{r}
.plotBias <- function(D,bias)
  # Plotting function to show gene level CpG density bias
  # Belinda Phipson
  # 5 March 2015
{
  o <- order(bias)
  splitf <- rep(1:100,each=200)[1:length(bias)]
  avgbias <- tapply(bias[o],factor(splitf),mean)
  sumDM <- tapply(D[o],factor(splitf),sum)
  propDM <- sumDM/table(splitf)
  graphics::par(mar=c(5,5,2,2))
  graphics::plot(avgbias,as.vector(propDM),
                 xlab="Length of gene (binned)",
                 ylab="Proportion Differential Methylation",cex.lab=1.5,
                 cex.axis=1.2)
  graphics::lines(stats::lowess(avgbias,propDM),col=4,lwd=2)
}
```

```{r}
txdb_genes_mod <- find_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         start = ifelse(strand == "+", start - 1500, start),
         end = ifelse(strand == "+", end, end + 1500)) %>% .[,c(1:4,6)] %>% as_granges(), 
  as_granges(wgbs_counts[,1:4])) %>% data.frame() %>%
  group_by(gene_id, seqnames, start, end, strand, width) %>%
  summarise(num_cg = n()) %>% ungroup() %>% data.frame()

gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(txdb_genes_mod)) %>% data.frame()
gene_info <- txdb_genes_mod %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)

.plotBias(D = gene_info$is_de, bias = as.vector(gene_info$width))
```

```{r}
txdb_genes_mod <- find_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         start = ifelse(strand == "+", start - 2000, end - 2000),
         end = ifelse(strand == "+", start + 4000, end + 2000)) %>% .[,c(1:4,6)] %>% as_granges(), 
  as_granges(wgbs_counts[,1:4])) %>% data.frame() %>%
  group_by(gene_id, seqnames, start, end, strand, width) %>%
  summarise(num_cg = n()) %>% ungroup() %>% data.frame()

gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(txdb_genes_mod)) %>% data.frame()
gene_info <- txdb_genes_mod %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)

.plotBias(D = gene_info$is_de, bias = as.vector(gene_info$width))
```
## run gst for these new bias param: 2k up/down

```{r}
txdb_genes_mod <- find_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         start = ifelse(strand == "+", start - 2000, end - 2000),
         end = ifelse(strand == "+", start + 4000, end + 2000)) %>% .[,c(1:4,6)] %>% as_granges(), 
  as_granges(wgbs_counts[,1:4])) %>% data.frame() %>%
  group_by(gene_id, seqnames, start, end, strand, width) %>%
  summarise(num_cg = n()) %>% ungroup() %>% data.frame()
gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(txdb_genes_mod)) %>% data.frame()
gene_info <- txdb_genes_mod %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)
```

```{r}
run_test <- list(
  sig.eg = gene_info %>% filter(is_de == 1) %>% .$gene_id,
  universe = gene_info$gene_id,
  freq = gene_info$num_cg,
  de = gene_info$is_de,
  equiv = gene_info$num_cg
)
```

```{r}
gst_2ud <- run_miss_methyl_b(run_test, method = "Wallenius")
gst_2ud_fisher <- run_miss_methyl_b(run_test, method = "Fishers")
gst_2ud_hyper <- run_miss_methyl_b(run_test, method = "None")
```

```{r}
gst_2ud <- gst_2ud[[1]][[1]]
gst_2ud_hyper <- gst_2ud_hyper[[1]][[1]]
gst_2ud$rank_fdr <- 1:nrow(gst_2ud)
gst_2ud_hyper$rank_fdr <- 1:nrow(gst_2ud_hyper)
```
```{r}
upset_input <- list(GOMethSeq = gst_1k %>% filter(FDR < 0.05) %>% .$GOID,
                    Biased = gst_1k_hyper %>% filter(FDR < 0.05) %>% .$GOID,
                    TSS2k_GOMethSeq = gst_2ud %>% filter(FDR < 0.05) %>% .$GOID,
                    TSS2k_Biased = gst_2ud_hyper %>% filter(FDR < 0.05) %>% .$GOID)
upset(fromList(upset_input), order.by = "freq", text.scale = 2)
```

```{r}
gst_2ud %>% filter(FDR < 0.05)
gst_2ud_hyper %>% filter(FDR < 0.05)
```

```{r}
go_entrez <- go_entrez %>%
  mutate(
         tss2k_gomethseq = ifelse(go %in% filter(gst_2ud, FDR < 0.05)$GOID, TRUE, FALSE),
         tss2k_biased = ifelse(go %in% filter(gst_2ud_hyper, FDR < 0.05)$GOID, TRUE, FALSE),
         rank_tss2k_gomethseq = gst_2ud[match(.$go, gst_2ud$GOID), "rank_fdr"],
         rank_tss2k_biased = gst_2ud_hyper[match(.$go, gst_2ud_hyper$GOID), "rank_fdr"])
```


```{r}
go_entrez %>% distinct(go, mean_num_cg, rank_tss2k_gomethseq, rank_tss2k_biased) %>%
  filter(rank_tss2k_gomethseq <= 1000, rank_tss2k_biased <= 1000) %>%
  ggplot(aes(x = mean_num_cg, y = rank_tss2k_biased - rank_tss2k_gomethseq)) + 
  geom_point() +
  geom_hline(yintercept = 0, colour = "red", linetype = "dashed") +
  labs(x = "Average No. CpGs in GO category", y = "Increase in rank in Wallenius") 
 # scale_x_continuous(limits = c(0, 1500))
```


```{r}
go_entrez %>% distinct(go, mean_num_cg, tss2k_gomethseq, tss2k_biased) %>%
  filter(tss2k_gomethseq != tss2k_biased) %>% 
  mutate(Method = ifelse(tss2k_gomethseq == FALSE, "Biased", "GOMethSeq")) %>%
  ggplot(aes(x = reorder(Method, mean_num_cg), y = mean_num_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(x = "Method", y = "Average number of CpGs per ontology category")

go_entrez %>% distinct(go, mean_num_cg, tss2k_gomethseq, tss2k_biased) %>%
  filter(tss2k_gomethseq == TRUE | tss2k_biased == TRUE) %>% 
  mutate(Method = case_when(tss2k_gomethseq == FALSE & tss2k_biased == TRUE ~ "Biased",
                            tss2k_gomethseq == TRUE & tss2k_biased == TRUE ~ "Both",
                            TRUE ~ "Bias Corrected")) %>%
  ggplot(aes(x = reorder(Method, mean_num_cg), y = mean_num_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(x = "Method", y = "Average number of CpGs per ontology category")
```


```{r}
truth_gst_100 <- rbind(
  gst_2ud[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_2ud_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")])

truth_gst_100 %>%
  ggplot(aes(x = rank_fdr, y = in_imm, colour = method)) +
  geom_line(size = 2) +
  labs(x = "Rank", y = "Cumulative no. in immune sets") +
  theme_bw()
```

* this didn't work
* but maybe if I show with gene length and it's only purpose is that?

# with gene length

```{r}
txdb_genes_mod <- find_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         start = ifelse(strand == "+", start - 2000, end - 2000),
         end = ifelse(strand == "+", start + 4000, end + 2000)) %>% .[,c(1:4,6)] %>% as_granges(), 
  as_granges(wgbs_counts[,1:4])) %>% data.frame() %>%
  group_by(gene_id, seqnames, start, end, strand, width) %>%
  summarise(num_cg = n()) %>% ungroup() %>% data.frame()
gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(txdb_genes_mod)) %>% data.frame()
gene_info <- txdb_genes_mod %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)
```

```{r}
run_test <- list(
  sig.eg = gene_info %>% filter(is_de == 1) %>% .$gene_id,
  universe = gene_info$gene_id,
  freq = gene_info$width,
  de = gene_info$is_de,
  equiv = gene_info$width
)
```

```{r}
gst_2ud_width <- run_miss_methyl_b(run_test, method = "Wallenius")
gst_2ud_width_fisher <- run_miss_methyl_b(run_test, method = "Fishers")
gst_2ud_width_hyper <- run_miss_methyl_b(run_test, method = "None")
```
```{r}
gst_2u
```
```{r}
1+1
```


###############

## around the TSS

```{r}
genes_with_pos <- pair_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         actual_start = start,
         actual_end = end,
         start = ifelse(strand == "+", start - 1500, end - 1500),
         end = ifelse(strand == "+", actual_start + 1500, end + 1500),
         actual_width = width) %>% .[,c(1:4,6:11)] %>% as_granges(), as_granges(wgbs_counts[,1:4])) %>% data.frame()
```

```{r}
genes_with_pos %>% mutate(cpg_pos = granges.y.start, dist_from_tss = cpg_pos - TSS) %>%
  group_by(dist_from_tss) %>% 
  summarise(n = n())
```

```{r}
genes_with_pos %>% mutate(cpg_pos = granges.y.start, dist_from_tss = cpg_pos - TSS) %>%
  ggplot(aes(x = dist_from_tss)) +
  geom_density()
```
```{r}
genes_with_pos %>% mutate(cpg_pos = granges.y.start, dist_from_tss = cpg_pos - TSS) %>%
  ggplot(aes(x = dist_from_tss)) +
  geom_bar()
```
```{r}
genes_with_pos %>% mutate(cpg_pos = granges.y.start, 
                          dist_from_tss = ifelse(granges.x.strand == "+", cpg_pos - TSS, TSS - cpg_pos)) %>%
  ggplot(aes(x = dist_from_tss)) +
  geom_bar()
```
* making this strand inclusive: if on positive strand cpg_pos - TSS: if cpg > TSS cpg downstream - will be on right of plot
* but if I kept it like that for both strands, then negative cpg > TSS means cpg upstream - but visually will be downstream
* so flip it so TSS > cpg - downstream

```{r}
genes_with_pos2 <- pair_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         actual_start = start,
         actual_end = end,
         start = ifelse(strand == "+", start - 1500, start),
         end = ifelse(strand == "+", end, end + 1500),
         actual_width = width) %>% .[,c(1:4,6:11)] %>% as_granges(), as_granges(wgbs_counts[,1:4])) %>% data.frame()
```

```{r}
genes_with_pos2 %>% mutate(cpg_pos = granges.y.start, dist_from_tss = cpg_pos - TSS) %>%
  ggplot(aes(x = dist_from_tss)) +
  geom_bar()
```
* ok something crazy skewed this
```{r}
genes_with_pos2 %>% mutate(cpg_pos = granges.y.start, 
                          dist_from_tss = ifelse(granges.x.strand == "+", cpg_pos - TSS, TSS - cpg_pos)) %>% summarise(max(dist_from_tss))
```
```{r}
txdb_genes %>%
  ggplot(aes(x = width)) +
  geom_density()

quantile(txdb_genes$width, probs = seq(0,1, by = 0.01))
```

```{r}
genes_with_pos2 %>% mutate(cpg_pos = granges.y.start, 
                          dist_from_tss = ifelse(granges.x.strand == "+", cpg_pos - TSS, TSS - cpg_pos)) %>% 
  filter(abs(dist_from_tss) <= 150000) %>%
  ggplot(aes(x = dist_from_tss)) +
  geom_bar()
```
```{r}
genes_with_pos2 %>% mutate(cpg_pos = granges.y.start, 
                          dist_from_tss = ifelse(granges.x.strand == "+", cpg_pos - TSS, TSS - cpg_pos)) %>% 
  filter(abs(dist_from_tss) <= 50000) %>%
  ggplot(aes(x = dist_from_tss)) +
  geom_bar() 
genes_with_pos2 %>% mutate(cpg_pos = granges.y.start, 
                          dist_from_tss = ifelse(granges.x.strand == "+", cpg_pos - TSS, TSS - cpg_pos)) %>% 
  filter(abs(dist_from_tss) <= 20000) %>%
  ggplot(aes(x = dist_from_tss)) +
  geom_bar() 
```

```{r}
genes_with_pos3 <- pair_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         actual_start = start,
         actual_end = end,
         start = ifelse(strand == "+", start - 5000, start),
         end = ifelse(strand == "+", end, end + 5000),
         actual_width = width) %>% .[,c(1:4,6:11)] %>% as_granges(), as_granges(wgbs_counts[,1:4])) %>% data.frame()

genes_with_pos3 %>% mutate(cpg_pos = granges.y.start, 
                          dist_from_tss = ifelse(granges.x.strand == "+", cpg_pos - TSS, TSS - cpg_pos)) %>% 
  filter(abs(dist_from_tss) <= 20000) %>%
  ggplot(aes(x = dist_from_tss)) +
  geom_bar()

genes_with_pos3 %>% mutate(cpg_pos = granges.y.start, 
                          dist_from_tss = ifelse(granges.x.strand == "+", cpg_pos - TSS, TSS - cpg_pos)) %>% 
  filter(abs(dist_from_tss) <= 10000) %>%
  ggplot(aes(x = dist_from_tss)) +
  geom_bar()
```

```{r}
genes_with_pos3 <- pair_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         actual_start = start,
         actual_end = end,
         start = ifelse(strand == "+", start - 10000, start),
         end = ifelse(strand == "+", end, end + 10000),
         actual_width = width) %>% .[,c(1:4,6:11)] %>% as_granges(), as_granges(wgbs_counts[,1:4])) %>% data.frame()

genes_with_pos3 %>% mutate(cpg_pos = granges.y.start, 
                          dist_from_tss = ifelse(granges.x.strand == "+", cpg_pos - TSS, TSS - cpg_pos)) %>% 
  filter(abs(dist_from_tss) <= 20000) %>%
  ggplot(aes(x = dist_from_tss)) +
  geom_bar()

genes_with_pos3 %>% mutate(cpg_pos = granges.y.start, 
                          dist_from_tss = ifelse(granges.x.strand == "+", cpg_pos - TSS, TSS - cpg_pos)) %>% 
  filter(abs(dist_from_tss) <= 10000) %>%
  ggplot(aes(x = dist_from_tss)) +
  geom_bar()
```

```{r}
genes_with_pos3 <- pair_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         actual_start = start,
         actual_end = end,
         start = ifelse(strand == "+", start - 2000, start),
         end = ifelse(strand == "+", end, end + 2000),
         actual_width = width) %>% .[,c(1:4,6:11)] %>% as_granges(), as_granges(wgbs_counts[,1:4])) %>% data.frame()

genes_with_pos3 %>% mutate(cpg_pos = granges.y.start, 
                          dist_from_tss = ifelse(granges.x.strand == "+", cpg_pos - TSS, TSS - cpg_pos)) %>% 
  filter(abs(dist_from_tss) <= 2000) %>%
  ggplot(aes(x = dist_from_tss)) +
  geom_bar()
```

```{r}
genes_with_pos3 <- pair_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         actual_start = start,
         actual_end = end,
         start = ifelse(strand == "+", start - 2000, start),
         end = ifelse(strand == "+", end, end + 2000),
         actual_width = width) %>% .[,c(1:4,6:11)] %>% as_granges(), as_granges(wgbs_counts[,1:4])) %>% data.frame()

genes_with_pos3 %>% mutate(cpg_pos = granges.y.start, 
                          dist_from_tss = ifelse(granges.x.strand == "+", cpg_pos - TSS, TSS - cpg_pos)) %>% 
  filter(abs(dist_from_tss) <= 2000) %>%
  ggplot(aes(x = dist_from_tss)) +
  geom_bar()
```


```{r}
genes_with_pos3 <- pair_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         actual_start = start,
         actual_end = end,
         start = ifelse(strand == "+", start - 2000, start),
         end = ifelse(strand == "+", end, end + 2000),
         actual_width = width) %>% .[,c(1:4,6:11)] %>% as_granges(), as_granges(wgbs_counts[,1:4])) %>% data.frame()
```



```{r}
genes_with_pos3 <- pair_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         actual_start = start,
         actual_end = end,
         start = ifelse(strand == "+", start - 20000, start),
         end = ifelse(strand == "+", end, end + 20000),
         actual_width = width) %>% .[,c(1:4,6:11)] %>% as_granges(), as_granges(wgbs_counts[,1:4])) %>% data.frame()

genes_with_pos3 %>% mutate(cpg_pos = granges.y.start, 
                          dist_from_tss = ifelse(granges.x.strand == "+", cpg_pos - TSS, TSS - cpg_pos)) %>% 
  filter(abs(dist_from_tss) <= 20000) %>%
  ggplot(aes(x = dist_from_tss)) +
  geom_bar()
```


# trying the percentile plot thing

```{r}
txdb_genes
```

```{r}
genes_with_pos2 %>% filter(gene_id == 1) %>% mutate(scale_100 = granges.x.width/100)
```

```{r}
genes_with_pos2 %>% filter(gene_id == 1) %>% mutate(percent_rank())
```

```{r}
?percent_rank
```

```{r}
genes_with_pos2 %>% filter(gene_id == 1) %>% mutate(try_rank = percent_rank(granges.y.start))
```
```{r}
genes_with_pos2 %>% filter(gene_id == 1) %>% mutate(try_rank = ntile(x = granges.y.start, 100))
```
* this version because it groups them easier
* other one is too exact, will be harder to compare across groups (genes)

```{r}
genes_with_pos2 %>%
  group_by(gene_id) %>%
  mutate(percent_rank = ntile(granges.y.start, 100)) %>%
  ggplot(aes(x = percent_rank)) +
  geom_density()
```
* gross

```{r}
genes_with_pos2 %>%
  group_by(gene_id) %>%
  mutate(percent_rank = ntile(granges.y.start, 100)) %>%
  ggplot(aes(x = percent_rank)) +
  geom_bar()
```
* also fairly gross

* maybe just gene length?
```{r}
pair_overlaps(txdb_genes %>% as_granges(), as_granges(wgbs_counts[,1:4])) %>% data.frame() %>%
  group_by(gene_id) %>%
  mutate(percent_rank = ntile(granges.y.start, 100)) %>%
  ggplot(aes(x = percent_rank)) +
  geom_bar()
```


* wait I'm not doing a density plot correctly
```{r}
genes_with_pos2 %>%
  group_by(gene_id) %>%
  mutate(percent_rank = ntile(granges.y.start, 100)) %>%
  group_by(gene_id, percent_rank) %>%
  mutate(gene_percent_length = max(granges.y.start) - min(granges.y.start),
         density_gene_percent = n()/gene_percent_length * 100) %>%
  ungroup() %>%
  distinct(gene_id, percent_rank, gene_percent_length, density_gene_percent) %>%
  ggplot(aes(x = percent_rank)) +
  geom_bar()
```
* ok still not right

```{r}
genes_with_pos2 %>%
  group_by(gene_id) %>%
  mutate(percent_rank = ntile(granges.y.start, 100)) %>%
  group_by(gene_id, percent_rank) %>%
  mutate(gene_percent_length = max(granges.y.start) - min(granges.y.start),
         density_gene_percent = n()/gene_percent_length * 100) %>%
  ungroup() %>%
  distinct(gene_id, percent_rank, gene_percent_length, density_gene_percent)
```

```{r}
genes_with_pos2 %>%
  group_by(gene_id) %>%
  mutate(percent_rank = ntile(granges.y.start, 100)) %>%
  group_by(gene_id, percent_rank) %>%
  mutate(gene_percent_length = max(granges.y.start) - min(granges.y.start)) %>%
  ungroup() %>% group_by(percent_rank) %>%
  mutate(density_gene_percent = n()/sum(gene_percent_length) * 100) %>%
  ungroup() %>%
  distinct(percent_rank, density_gene_percent) %>% mutate(sum(density_gene_percent))
  ggplot(aes(x = ))
```
* not working - leave it alone

* maybe a boxplot?

```{r}
genes_with_pos2 %>%
  group_by(gene_id) %>%
  mutate(percent_rank = ntile(granges.y.start, 100)) %>%
  ungroup() %>% group_by(percent_rank) %>%
  summarise(n=n()) %>%
  ggplot(aes(x = as.factor(percent_rank), y = n)) +
  geom_boxplot()
```

### more than 2000 bp - because I think that was too narrow so would just work anyway

```{r}
txdb_genes_mod <- find_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end),
         start = ifelse(strand == "+", start - 5000, end - 5000),
         end = ifelse(strand == "+", start + 10000, end + 5000)) %>% .[,c(1:4,6)] %>% as_granges(), 
  as_granges(wgbs_counts[,1:4])) %>% data.frame() %>%
  group_by(gene_id, seqnames, start, end, strand, width) %>%
  summarise(num_cg = n()) %>% ungroup() %>% data.frame()
gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(txdb_genes_mod)) %>% data.frame()
gene_info <- txdb_genes_mod %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)
```

```{r}
run_test <- list(
  sig.eg = gene_info %>% filter(is_de == 1) %>% .$gene_id,
  universe = gene_info$gene_id,
  freq = gene_info$num_cg,
  de = gene_info$is_de,
  equiv = gene_info$num_cg
)
```

```{r}
gst_5ud <- run_miss_methyl_b(run_test, method = "Wallenius")
#gst_2ud_fisher <- run_miss_methyl_b(run_test, method = "Fishers")
gst_5ud_hyper <- run_miss_methyl_b(run_test, method = "None")
```

```{r}
run_test <- list(
  sig.eg = gene_info %>% filter(is_de == 1) %>% .$gene_id,
  universe = gene_info$gene_id,
  freq = gene_info$width,
  de = gene_info$is_de,
  equiv = gene_info$width
)
```

```{r}
gst_5ud_width <- run_miss_methyl_b(run_test, method = "Wallenius")
#gst_2ud_width_fisher <- run_miss_methyl_b(run_test, method = "Fishers")
gst_5ud_width_hyper <- run_miss_methyl_b(run_test, method = "None")
```

```{r}
gst_5ud <- gst_5ud[[1]]
gst_5ud_hyper <- gst_5ud_hyper[[1]]
gst_5ud$rank_fdr <- 1:nrow(gst_5ud)
gst_5ud_hyper$rank_fdr <- 1:nrow(gst_5ud_hyper)
```


```{r}
gst_5ud_width <- gst_5ud_width[[1]]
gst_5ud_width_hyper <- gst_5ud_width_hyper[[1]]
gst_5ud_width$rank_fdr <- 1:nrow(gst_5ud_width)
gst_5ud_width_hyper$rank_fdr <- 1:nrow(gst_5ud_width_hyper)
```

```{r}
upset_input <- list(GOMethSeq = gst_1k %>% filter(FDR < 0.05) %>% .$GOID,
                    Biased = gst_1k_hyper %>% filter(FDR < 0.05) %>% .$GOID,
                    TSS5k_GOMethSeq = gst_5ud %>% filter(FDR < 0.05) %>% .$GOID,
                    TSS5k_Biased = gst_5ud_hyper %>% filter(FDR < 0.05) %>% .$GOID,
                    TSS5k_width_GOMethSeq = gst_5ud_width %>% filter(FDR < 0.05) %>% .$GOID,
                    TSS5k_width_Biased = gst_5ud_width_hyper %>% filter(FDR < 0.05) %>% .$GOID)
upset(fromList(upset_input), order.by = "freq", text.scale = 2)
```

```{r}
truth_gst_100 <- rbind(
  gst_1k[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_1k_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_2ud[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS2k_GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_2ud_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS2k_Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")])

truth_gst_100 %>%
  ggplot(aes(x = rank_fdr, y = in_imm, colour = method)) +
  geom_line() +
  labs(x = "Rank", y = "Cumulative no. in immune sets") +
  theme_bw()
```


```{r}
truth_gst_100 <- rbind(
  gst_5ud[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS5k_GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_5ud_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS5k_Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")])

truth_gst_100 %>%
  ggplot(aes(x = rank_fdr, y = in_imm, colour = method)) +
  geom_line() +
  labs(x = "Rank", y = "Cumulative no. in immune sets") +
  theme_bw()
```

```{r}
truth_gst_100 <- rbind(
  gst_5ud_width[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS5k_width_GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_5ud_width_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS5k_with_Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")])

truth_gst_100 %>%
  ggplot(aes(x = rank_fdr, y = in_imm, colour = method)) +
  geom_line() +
  labs(x = "Rank", y = "Cumulative no. in immune sets") +
  theme_bw()
```

```{r}
truth_gst_100 <- rbind(
  gst_5ud[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS5k_GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_5ud_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS5k_Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_5ud_width[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS5k_width_GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_5ud_width_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS5k_width_Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")])

truth_gst_100 %>%
  ggplot(aes(x = rank_fdr, y = in_imm, colour = method)) +
  geom_line() +
  labs(x = "Rank", y = "Cumulative no. in immune sets") +
  theme_bw()
```


```{r}
truth_gst_100 <- rbind(
  gst_5ud[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS5k_GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_5ud_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS5k_Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")])

truth_gst_100 %>%
  ggplot(aes(x = rank_fdr, y = in_imm, colour = method)) +
  geom_line() +
  labs(x = "Rank", y = "Cumulative no. in immune sets") +
  theme_bw()
```


```{r}
truth_gst_100 <- rbind(
  gst_2ud[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS2k_GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_2ud_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS2k_Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")])

truth_gst_100 %>%
  ggplot(aes(x = rank_fdr, y = in_imm, colour = method)) +
  geom_line() +
  labs(x = "Rank", y = "Cumulative no. in immune sets") +
  theme_bw()
```

# just original widths

```{r}
txdb_genes_mod <- find_overlaps(txdb_genes %>%
  mutate(TSS = ifelse(strand == "+", start, end)) %>% .[,c(1:4,6)] %>% as_granges(), 
  as_granges(wgbs_counts[,1:4])) %>% data.frame() %>%
  group_by(gene_id, seqnames, start, end, strand, width) %>%
  summarise(num_cg = n()) %>% ungroup() %>% data.frame()
gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr[1:1000,], id = 1:1000)), as_granges(txdb_genes_mod)) %>% data.frame()
gene_info <- txdb_genes_mod %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)
```

```{r}
run_test <- list(
  sig.eg = gene_info %>% filter(is_de == 1) %>% .$gene_id,
  universe = gene_info$gene_id,
  freq = gene_info$width,
  de = gene_info$is_de,
  equiv = gene_info$width
)
```

```{r}
gst_width <- run_miss_methyl_b(run_test, method = "Wallenius")
#gst_2ud_fisher <- run_miss_methyl_b(run_test, method = "Fishers")
gst_width_hyper <- run_miss_methyl_b(run_test, method = "None")
```

```{r}
gst_width <- gst_width[[1]]
gst_width_hyper <- gst_width_hyper[[1]]
gst_width$rank_fdr <- 1:nrow(gst_width)
gst_width_hyper$rank_fdr <- 1:nrow(gst_width_hyper)
```

```{r}
truth_gst_100 <- rbind(
  gst_width[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "width_GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_width_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "width_Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")])

truth_gst_100 %>%
  ggplot(aes(x = rank_fdr, y = in_imm, colour = method)) +
  geom_line() +
  labs(x = "Rank", y = "Cumulative no. in immune sets") +
  theme_bw()
```



```{r}
txdb_genes_mod <- find_overlaps(txdb_genes %>%
  mutate(original_width = width,
        TSS = ifelse(strand == "+", start, end),
         start = ifelse(strand == "+", start - 2000, end - 2000),
         end = ifelse(strand == "+", start + 4000, end + 2000)) %>% .[,c(1:4,6,7,8)] %>% as_granges(), 
  as_granges(wgbs_counts[,1:4])) %>% data.frame() %>%
  group_by(gene_id, seqnames, start, end, strand, width, num_cg, original_width) %>%
  summarise(tss_2k_num_cg = n(),
            width_2k = width) %>% ungroup() %>% data.frame()
txdb_genes_mod <- distinct(txdb_genes_mod)
```


```{r}
txdb_genes_mod %>%
  .[,c(1,7:10)] %>%
  melt("gene_id") %>%
  ggplot(aes(x = log2(value), colour = variable)) +
  geom_density() +
  theme_bw()

txdb_genes_mod %>%
  .[,c(1,7:10)] %>%
  melt("gene_id") %>%
  ggplot(aes(x = log2(value), colour = variable, fill = variable)) +
  geom_histogram(alpha = 0.4) +
  theme_bw()
```


```{r}
txdb_genes_mod %>%
  .[,c(1,6,7,8)] %>%
  melt("gene_id") %>%
  ggplot(aes(x = log2(value), colour = variable, fill = variable)) +
  geom_histogram() +
  theme_bw()
```


```{r}
txdb_genes_mod %>%
  ggplot(aes(x = tss_2k_num_cg)) +
  geom_density()
txdb_genes_mod %>%
  ggplot(aes(x = num_cg)) +
  geom_density()
```
```{r}
?melt()
```

```{r}
txdb_genes_mod <- find_overlaps(txdb_genes %>%
  mutate(original_width = width,
        TSS = ifelse(strand == "+", start, end),
         start = ifelse(strand == "+", start - 2000, end - 2000),
         end = ifelse(strand == "+", start + 4000, end + 2000)) %>% .[,c(1:4,6,7,8)] %>% as_granges(), 
  as_granges(wgbs_counts[,1:4])) %>% data.frame() %>%
  group_by(gene_id, seqnames, start, end, strand, width, num_cg, original_width) %>%
  summarise(tss_2k_num_cg = n(),
            width_2k = width) %>% ungroup() %>% data.frame()
txdb_genes_mod <- distinct(txdb_genes_mod)

gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(txdb_genes_mod)) %>% data.frame()
gene_info <- txdb_genes_mod %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width, tss_2k_num_cg, original_width)

.plotBias(D = gene_info$is_de, bias = as.vector(gene_info$tss_2k_num_cg))
```

```{r}
.plotBias(D = gene_info$is_de, bias = as.vector(gene_info$width))
```

```{r}
gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(txdb_genes)) %>% data.frame()
txdb_genes %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% group_by(is_de) %>% summarise(n=n()) %>%
  ungroup() %>% mutate(prop = n/sum(n))
```

```{r}
dmrcate_seq_anno <- readRDS("../output/dmrcate_seq_anno.rds")
dmrcate_seq_anno <- dmrcate_seq_anno@ranges %>% data.frame()
dmrcate_seq_anno <- dmrcate_seq_anno %>% filter(is.sig == TRUE)
```

```{r}
gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_anno, id = 1:nrow(dmrcate_seq_anno))), as_granges(txdb_genes)) %>% data.frame()
txdb_genes %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% group_by(is_de) %>% summarise(n=n()) %>%
  ungroup() %>% mutate(prop = n/sum(n))
```

```{r}
truth_gst_100 <- rbind(
  gst[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_2ud[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS2k_GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_2ud_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS2k_Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_2ud_width_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS2k_width_Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")])

truth_gst_100 %>%
  ggplot(aes(x = rank_fdr, y = in_imm, colour = method)) +
  geom_line() +
  labs(x = "Rank", y = "Cumulative no. in immune sets") +
  theme_bw()
```

```{r}
truth_gst_100 <- rbind(
  gst[1:200,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_hyper[1:200,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_2ud[1:200,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS2k_GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_2ud_hyper[1:200,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS2k_Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_2ud_width_hyper[1:200,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS2k_width_Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")])

truth_gst_100 %>%
  ggplot(aes(x = rank_fdr, y = in_imm, colour = method)) +
  geom_line() +
  labs(x = "Rank", y = "Cumulative no. in immune sets") +
  theme_bw()
```

```{r}
truth_gst_100 <- rbind(
 # gst_2ud[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS2k_GOMethSeq") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_2ud_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS2k_Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")],
  gst_2ud_width_hyper[1:100,] %>% mutate(in_imm = cumsum(GOID %in% immuneGO$GOID), method = "TSS2k_width_Biased") %>% .[names(.) %in% c("rank_fdr", "in_imm", "method")])

truth_gst_100 %>%
  ggplot(aes(x = rank_fdr, y = in_imm, colour = method)) +
  geom_line() +
  labs(x = "Rank", y = "Cumulative no. in immune sets") +
  theme_bw()
```


```{r}
go_entrez
```






```{r}
txdb_genes_mod <- find_overlaps(txdb_genes %>%
  mutate(original_width = width,
        TSS = ifelse(strand == "+", start, end),
         start = ifelse(strand == "+", start - 2000, end - 2000),
         end = ifelse(strand == "+", start + 4000, end + 2000)) %>% .[,c(1:4,6,7,8)] %>% as_granges(), 
  as_granges(wgbs_counts[,1:4])) %>% data.frame() %>%
  group_by(gene_id, seqnames, start, end, strand, width, num_cg, original_width) %>%
  summarise(tss_2k_num_cg = n(),
            width_2k = width) %>% ungroup() %>% data.frame()
txdb_genes_mod <- distinct(txdb_genes_mod)

gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(txdb_genes_mod)) %>% data.frame()
gene_info_2kb <- txdb_genes_mod %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width, tss_2k_num_cg, original_width)
```
```{r}
gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))), as_granges(txdb_genes)) %>% data.frame()
gene_info_og <- txdb_genes_mod %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)

gene_try <- find_overlaps(as_granges(mutate(dmrcate_seq_dmr, id = 1:nrow(dmrcate_seq_dmr))[1:1000,]), as_granges(txdb_genes)) %>% data.frame()
gene_info_og_1k <- txdb_genes_mod %>% mutate(is_de = ifelse(gene_id %in% gene_try$gene_id, 1, 0)) %>% distinct(gene_id, num_cg, is_de, width)
```

```{r}
upset_input <- list(gene_2kb = gene_info_2kb %>% filter(is_de == 1) %>% .$gene_id,
                    gene_og = gene_info_og %>% filter(is_de == 1) %>% .$gene_id,
                    gene_top1k = gene_info_og_1k %>% filter(is_de == 1) %>% .$gene_id)
upset(fromList(upset_input), order.by = "freq", text.scale = 2)
```

```{r}
upset_input <- list(gene_2kb = gene_info_2kb %>% filter(is_de == 1) %>% .$gene_id,
                    top1k_dmrs = gene_info_og_1k %>% filter(is_de == 1) %>% .$gene_id)
upset(fromList(upset_input), order.by = "freq", text.scale = 2)
```


```{r}
upset_input <- list(gene_2kb = gst_2ud %>% filter(FDR < 0.05) %>% .$GOID,
                    gene_2kb_width = gst_2ud_width %>% filter(FDR < 0.05) %>% .$GOID,
                    top1k_dmrs = gst_1k %>% filter(FDR < 0.05) %>% .$GOID)
upset(fromList(upset_input), order.by = "freq", text.scale = 2)
```




```{r}
gst_2ud
```


######



```{r}
go_entrez
```




```{r}
go_entrez <- go_entrez %>%
  mutate(
         tss2k_width_gomethseq = ifelse(go %in% filter(gst_2ud_width, FDR < 0.05)$GOID, TRUE, FALSE),
         tss2k_width_biased = ifelse(go %in% filter(gst_2ud_width_hyper, FDR < 0.05)$GOID, TRUE, FALSE),
         rank_tss2k_width_gomethseq = gst_2ud_width[match(.$go, gst_2ud_width$GOID), "rank_fdr"],
         rank_tss2k_width_biased = gst_2ud_width_hyper[match(.$go, gst_2ud_width_hyper$GOID), "rank_fdr"])
```


```{r}
go_entrez %>% distinct(go, mean_num_cg, rank_tss2k_gomethseq, rank_tss2k_biased) %>%
  filter(rank_tss2k_gomethseq <= 1000, rank_tss2k_biased <= 1000) %>%
  ggplot(aes(x = mean_num_cg, y = rank_tss2k_biased - rank_tss2k_gomethseq)) + 
  geom_point() +
  geom_hline(yintercept = 0, colour = "red", linetype = "dashed") +
  labs(x = "Average No. CpGs in GO category", y = "Increase in rank in Wallenius") 
 # scale_x_continuous(limits = c(0, 1500))
```


```{r}
go_entrez %>% distinct(go, mean_num_cg, tss2k_gomethseq, tss2k_biased) %>%
  filter(tss2k_gomethseq != tss2k_biased) %>% 
  mutate(Method = ifelse(tss2k_gomethseq == FALSE, "Biased", "GOMethSeq")) %>%
  ggplot(aes(x = reorder(Method, mean_num_cg), y = mean_num_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(x = "Method", y = "Average number of CpGs per ontology category")

go_entrez %>% distinct(go, mean_num_cg, tss2k_gomethseq, tss2k_biased) %>%
  filter(tss2k_gomethseq == TRUE | tss2k_biased == TRUE) %>% 
  mutate(Method = case_when(tss2k_gomethseq == FALSE & tss2k_biased == TRUE ~ "Biased",
                            tss2k_gomethseq == TRUE & tss2k_biased == TRUE ~ "Both",
                            TRUE ~ "Bias Corrected")) %>%
  ggplot(aes(x = reorder(Method, mean_num_cg), y = mean_num_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(x = "Method", y = "Average number of CpGs per ontology category")
```
