==> analysis/03_checkGSTInputs.Rmd <==
---
title: "03_checkGSTInputs"
author: "Caitlin Page"
date: "2025-02-27"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: inline
---

# can use dmrcate because have dmr for array
```{r}
dmrcate_array_anno <- readRDS("../output/dmrcate_array_anno.rds")
dmrcate_array_anno <- data.frame(dmrcate_array_anno@ranges)
```
```{r}
dmrcate_array_anno$pos <- paste0(dmrcate_array_anno$seqnames, "-", dmrcate_array_anno$start)
dmrcate_overlap <- find_overlaps(as_granges(dmrcate_array_dmr), as_granges(dmrcate_array_anno)) %>% data.frame()
```
```{r}
dmrcate_array_anno <- dmrcate_array_anno %>% mutate(in_dmr = ifelse(pos %in% dmrcate_overlap$pos, TRUE, FALSE))
```
```{r}
dmrcate_array_anno <- find_overlaps(as_granges(dmrcate_array_anno), as_granges(mutate(anno_array, cg_id = rownames(anno_array)))) %>% data.frame()
```

```{r}
mm_map_dmrcate_array <- getMappedEntrezIDs(filter(dmrcate_array_anno, in_dmr == TRUE)$cg_id, dmrcate_array_anno$cg_id, "EPIC")
mm_map_dmrcate_array_body <- getMappedEntrezIDs(filter(dmrcate_array_anno, in_dmr == TRUE)$cg_id, dmrcate_array_anno$cg_id, "EPIC", genomic.features = "Body")
```

```{r}
getMap
```

```{r}
dmrcate_array_bio_anno <- annoGeneDmr(anno_array, dmrcate_array_anno, gene_source = "biomart", biomart_genes)
```
* I can run - MM
* and just the gsa-meth bit using my input beforehand
* and see if results change
* big brain moment
* and can just run the mapped entrez bit
* and can really compare results
* dmrcate array stuff, and also just regular array
```{r}
me_gst_dmrcate_array_out <- run_miss_methyl_a(biomart_genes, annoGeneDmr(anno_array, dmrcate_array_anno, gene_source = "biomart", biomart_genes))
```
```{r}

```

# compare my out and mm out - theoretically should be the same
```{r}
mm_map_dmrcate_array$sig.eg %>% length()
mm_map_dmrcate_array_body$sig.eg %>% length()
me_gst_dmrcate_array_out$sig.eg %>% length()
```

```{r}
dmrcate_array_dmr
```
```{r}
dmrcate_array_anno
```


```{r}
ggVennDiagram(list(mm = mm_map_dmrcate_array$sig.eg, mm_body = mm_map_dmrcate_array_body$sig.eg, me = me_gst_dmrcate_array_out$sig.eg))
ggVennDiagram(list(mm = mm_map_dmrcate_array$universe, mm_body = mm_map_dmrcate_array_body$universe, me = me_gst_dmrcate_array_out$universe))
```
```{r}
anno_array
```

```{r}
mm_dmrcate_array <- gometh(filter(dmrcate_array_anno, in_dmr == TRUE)$cg_id, dmrcate_array_anno$cg_id, "GO", "EPIC")
mm_dmrcate_array %>% .[order(.$FDR),]
```
```{r}
mm_dmrcate_array2 <- goregion(as_granges(dmrcate_array_dmr[1:5000,]), dmrcate_array_anno$cg_id, "GO", "EPIC") %>% .[order(.$FDR),]
mm_dmrcate_array2_null <- goregion(as_granges(dmrcate_array_dmr[1:5000,]), dmrcate_array_anno$cg_id, "GO", "EPIC", prior.prob = FALSE) %>% .[order(.$FDR),]
```

```{r}
me_gst_dmrcate_array <- run_miss_methyl_b(me_gst_dmrcate_array_out, method = "Wallenius")
me_gst_dmrcate_array %>% .[order(.$FDR),]
```
* ok so good news is they both look not great
```{r}
dmrcate_array_anno
mm_dmrcate_array_indiv <- gometh(filter(dmrcate_array_anno, is.sig == TRUE)$cg_id, dmrcate_array_anno$cg_id, "GO", "EPIC")
mm_dmrcate_array_indiv %>% .[order(.$FDR),]
```
* wow ok this matches the other one
* oh probs needs less data
* hey maybe I need to be more restrictive in mine? - too much noise?
```{r}
dmrcate_array_anno <- dmrcate_array_anno %>% .[order(.$ind.fdr),] %>% mutate(rank = 1:n())
```
```{r}
length(filter(dmrcate_array_anno, is.sig == TRUE)$cg_id)
```

```{r}
gometh(filter(dmrcate_array_anno, rank <= 10000)$cg_id, dmrcate_array_anno$cg_id, "GO", "EPIC") %>% .[order(.$FDR),]
gometh(filter(dmrcate_array_anno, rank <= 100000)$cg_id, dmrcate_array_anno$cg_id, "GO", "EPIC") %>% .[order(.$FDR),]
```
* ok this is what its supposed to show
* 100 k - too many
* what about 50k
* 50k gives immune but more vague
```{r}
gometh(filter(dmrcate_array_anno, rank <= 50000)$cg_id, dmrcate_array_anno$cg_id, "GO", "EPIC") %>% .[order(.$FDR),]
```
```{r}
goregion(as_granges(dmrcate_array_dmr), dmrcate_array_anno$cg_id, "GO", "EPIC") %>% .[order(.$FDR),] #good me picking the sites gives same results
goregion(as_granges(dmrcate_array_dmr[1:5000,]), dmrcate_array_anno$cg_id, "GO", "EPIC") %>% .[order(.$FDR),]
goregion(as_granges(dmrcate_array_dmr[1:1000,]), dmrcate_array_anno$cg_id, "GO", "EPIC") %>% .[order(.$FDR),]
```
* filter the dmrs more - and get better results
* oh it would be ironic if this was what was preventing better terms for sequencing (and it probably is)
```{r}
goregion(as_granges(dmrcate_array_dmr[1:500,]), dmrcate_array_anno$cg_id, "GO", "EPIC") %>% .[order(.$FDR),]
```

```{r}
dmrcate_array_dmr
```

```{r}
DMPs
gometh(filter(DMPs, adj.P.Val < 0.05)$cg_id, DMPs$cg_id, "GO", "EPIC") %>% .[order(.$FDR),]
```
```{r}
gometh(filter(DMPs, rank <= 10000)$cg_id, DMPs$cg_id, "GO", "EPIC") %>% .[order(.$FDR),]
```
* me again
```{r}
run_miss_methyl_b(run_miss_methyl_a(biomart_genes, annoGeneDmr(anno_array, filter(dmrcate_array_anno, rank <= 10000), gene_source = "biomart", biomart_genes)), method = "Wallenius")
```
* still no good
```{r}
.plotBias(me_gst_dmrcate_array_out$de, as.vector(me_gst_dmrcate_array_out$freq))
```
* this is too high b/c it's using the seq way of doing things
```{r}
gometh(filter(dmrcate_array_anno, rank <= 10000)$cg_id, dmrcate_array_anno$cg_id, "GO", "EPIC", plot.bias = TRUE)
```

```{r}
gometh(filter(DMPs, rank <= 10000)$cg_id, DMPs$cg_id, "GO", "EPIC", plot.bias = TRUE) 
```
```{r}
me_mm_dmrcate_array <- mm_cheat(me_gst_dmrcate_array_out)
```
```{r}
me_mm_dmrcate_array %>% .[order(.$FDR),]
```
```{r}
biomart_genes_array <- find_overlaps(as_granges(biomart_genes), as_granges(dmrcate_array_anno)) %>% data.frame() %>% group_by(ensembl_gene_id) %>% mutate(num_cg = n()) %>% ungroup() %>% .[,c(1:14)] %>% distinct(.)
```
```{r}
biomart_genes
```
```{r}
run_miss_methyl_a(biomart_genes, annoGeneDmr(anno_array, dmrcate_array_anno, gene_source = "biomart", biomart_genes))
```

```{r}
unique(biomart_genes$ensembl_gene_id) %>% length()
```

```{r}
mm_cheat <- function(out, plot.bias = TRUE, prior.prob = TRUE, equiv.cpg=FALSE, sig.genes = FALSE, fract.counts = FALSE) {
  go <- .getGO()
  collection <- go$idList
  sorted.eg.sig <- out$sig.eg
  eg.universe <- out$universe
  freq_genes <- out$freq
  test.de <- out$de
  #frac <- out$fract.counts
  equiv <- out$equiv
  
  # Check collection is a list with character vectors
  if(!is.list(collection))
    collection <- list(collection=collection)
  collection <- lapply(collection, as.character)
  # Make sure gene set collections don't have any NAs
  collection <- lapply(collection, function(x) x[!is.na(x)])
  # Remove genes that are NOT in the universe from collections
  collection <- lapply(collection, function(x) x[x %in% eg.universe])
  # Remove collections with no genes left after universe filter
  inUniv <- sapply(collection, function(x) length(x) > 0)
  collection <- collection[inUniv]

  # Estimate prior probabilities
  if(prior.prob){
    if(equiv.cpg){ 
        # use "equivalent" no. of cpgs in odds calculation
        pwf <- .estimatePWF(D=test.de,bias=as.vector(equiv))
        if(plot.bias)
            .plotBias(D=test.de,bias=as.vector(equiv))
    } else {
        pwf <- .estimatePWF(D=test.de,bias=as.vector(freq_genes))
        if(plot.bias)
            .plotBias(D=test.de,bias=as.vector(freq_genes))
        
    } 
  }
  
  results <- matrix(NA,ncol=4,nrow=length(collection))
  colnames(results) <- c("N","DE","P.DE","FDR")
  rownames(results) <- names(collection)
  results[,"N"] <- unlist(lapply(collection,length))
  if(sig.genes) SigGenesInSet <- rep(NA,length(collection))
  
  if(prior.prob & fract.counts){ 
      # use fractional counting to account for cpgs that map to multiple genes
      results[,"DE"] <- unlist(lapply(collection, function(x) 
          sum((sorted.eg.sig %in% x) * frac$frac)))
  } else {
      results[,"DE"] <- unlist(lapply(collection, function(x) 
          sum((sorted.eg.sig %in% x))))
  } 
  
  Nuniverse <- length(eg.universe)
  m <- length(sorted.eg.sig)
  
  # Hypergeometric test with prior probabilities
  if(prior.prob){
    for(i in 1:length(collection)){
      InSet <- eg.universe %in% collection[[i]]
      pw.red <- sum(pwf[InSet])/results[i,"N"]
      pw.white <- sum(pwf[!InSet])/(Nuniverse-results[i,"N"])
      odds <- pw.red/pw.white
      results[i,"P.DE"] <- BiasedUrn::pWNCHypergeo(results[i,"DE"],
                                                   results[i,"N"],
                                                   Nuniverse-results[i,"N"],
                                                   m,odds,lower.tail=FALSE) + 
          BiasedUrn::dWNCHypergeo(results[i,"DE"],
                                  results[i,"N"],
                                  Nuniverse-results[i,"N"],
                                  m,odds)
      if(sig.genes){
        # Get gene symbols of significant genes
        SigGenesEntrezID <- sorted.eg.sig[sorted.eg.sig %in% collection[[i]]]
        SigGenesSymbol <- suppressMessages(AnnotationDbi::select(org.Hs.eg.db, 
                                             keys = SigGenesEntrezID,
                                             columns = "SYMBOL"))
        SigGenesInSet[i] <- paste(SigGenesSymbol$SYMBOL,collapse=",")
      }
      if(results[i,"P.DE"]==0){
        message("Pvalue of exactly zero detected. Performing hypergeometric 
                test for gene set ", rownames(results)[i])
        results[i,"P.DE"] <- stats::phyper(q=results[i,"DE"]-0.5,m=m,
                                           n=Nuniverse-m,k=results[i,"N"],
                                           lower.tail=FALSE)
      }
    }
  }
  # Hypergeometric test without prior probabilities
  else{
    for(i in 1:length(collection)){
      results[i,"P.DE"] <- stats::phyper(q=results[i,"DE"]-0.5,m=m,
                                         n=Nuniverse-m,k=results[i,"N"],
                                         lower.tail=FALSE)
      if(sig.genes){
        # Get gene symbols of significant genes
        SigGenesEntrezID <- sorted.eg.sig[sorted.eg.sig %in% collection[[i]]]
        SigGenesSymbol <- suppressMessages(AnnotationDbi::select(org.Hs.eg.db, 
                                                               keys = SigGenesEntrezID,
                                                               columns = "SYMBOL"))
        SigGenesInSet[i] <- paste(SigGenesSymbol$SYMBOL,collapse=",")
      }
    }
  }
  results[,"FDR"] <- stats::p.adjust(results[,"P.DE"],method="BH")
  results[,"DE"] <- floor(results[,"DE"])
  if(sig.genes) data.frame(results, SigGenesInSet)
  else data.frame(results)
  results <- merge(go$idTable,results,by.x="GOID",by.y="row.names")
    rownames(results) <- results$GOID
    results <- results %>% .[order(.$FDR),] %>% mutate(rank_fdr = 1:n())
    results
}


```

```{r}
dmrcate_array_dmr
dmrcate_array_anno
```

```{r}
mm_cheat_plot <- function(out, group_size = c(200,100)) {
  go <- .getGO()
  collection <- go$idList
  sorted.eg.sig <- out$sig.eg
  eg.universe <- out$universe
  freq_genes <- out$freq
  test.de <- out$de
  #frac <- out$fract.counts
  equiv <- out$equiv
  
  # Check collection is a list with character vectors
  if(!is.list(collection))
    collection <- list(collection=collection)
  collection <- lapply(collection, as.character)
  # Make sure gene set collections don't have any NAs
  collection <- lapply(collection, function(x) x[!is.na(x)])
  # Remove genes that are NOT in the universe from collections
  collection <- lapply(collection, function(x) x[x %in% eg.universe])
  # Remove collections with no genes left after universe filter
  inUniv <- sapply(collection, function(x) length(x) > 0)
  collection <- collection[inUniv]

  # Estimate prior probabilities
        # use "equivalent" no. of cpgs in odds calculation
  pwf <- .estimatePWF(D=test.de,bias=as.vector(equiv))
  .plotBias(D=test.de,bias=as.vector(freq_genes), group_size = group_size)

}
```

```{r}
mm_array_res
```


.getGO <- function(){
  if(!requireNamespace("org.Hs.eg.db", quietly = TRUE))
    stop("org.Hs.eg.db package required but not installed.")
  egGO2ALLEGS <- utils::getFromNamespace("org.Hs.egGO2ALLEGS", "org.Hs.eg.db")
  GeneID.PathID <- AnnotationDbi::toTable(egGO2ALLEGS)[,c("gene_id", "go_id", "Ontology")]
  d <- !duplicated(GeneID.PathID[, c("gene_id", "go_id")])
  GeneID.PathID <- GeneID.PathID[d, ]
  GOID.TERM <- suppressMessages(AnnotationDbi::select(GO.db::GO.db,
                                                      keys=unique(GeneID.PathID$go_id),
                                                      columns=c("GOID","ONTOLOGY","TERM"),
                                                      keytype="GOID"))
  go <- tapply(GeneID.PathID$gene_id, GeneID.PathID$go_id, list)

  list(idList=go, idTable=GOID.TERM)
}


.estimatePWF <- function(D,bias)
  # An alternative to goseq function nullp, which is transformation invariant
  # Belinda Phipson and Gordon Smyth
  # 6 March 2015
{
  prior.prob <- bias
  o <- order(bias)
  prior.prob[o] <- limma::tricubeMovingAverage(D[o],span=0.5)
  prior.prob
}
test_pwf <- .estimatePWF(D=test.de,bias=as.vector(freq_genes))

.plotBias <- function(D,bias)
  # Plotting function to show gene level CpG density bias
  # Belinda Phipson
  # 5 March 2015
{
  o <- order(bias)
  splitf <- rep(1:100,each=200)[1:length(bias)]
  avgbias <- tapply(bias[o],factor(splitf),mean)
  sumDM <- tapply(D[o],factor(splitf),sum)
  propDM <- sumDM/table(splitf)
  graphics::par(mar=c(5,5,2,2))
  graphics::plot(avgbias,as.vector(propDM),
                 xlab="Number of CpGs per gene (binned)",
                 ylab="Proportion Differential Methylation",cex.lab=1.5,
                 cex.axis=1.2)
  graphics::lines(stats::lowess(avgbias,propDM),col=4,lwd=2)
}
* same input as .estimatePWF

# get the genes and dmrs organised
run_miss_methyl_a <- function(biomart_genes, test_anno) { # add something about extending the bias to include upstream of gene maybe??
  #get entrez gene ids and gene symbols
  entrez_genes <- AnnotationDbi::select(org.Hs.eg.db,
                                        keys=AnnotationDbi::keys(org.Hs.eg.db),
                                        columns=c("ENTREZID","SYMBOL"),
                                        keytype="ENTREZID")
  biomart_genes$entrez_id <- entrez_genes[match(biomart_genes$gene_name, entrez_genes$SYMBOL), "ENTREZID"]
  biomart_genes_filt <- biomart_genes[!biomart_genes$entrez_id %in% NA,]
  biomart_genes_filt <- biomart_genes_filt %>% group_by(gene_name) %>% filter(num_cg == max(num_cg)) %>% ungroup() #this line will be slow

  test_anno$entrez_id <- entrez_genes[match(test_anno$gene_name, entrez_genes$SYMBOL), "ENTREZID"]
  test_anno_filt <- test_anno[!test_anno$entrez_id %in% NA,]
  test_anno_filt <- test_anno_filt[test_anno_filt$tss_rel_peak %in% "Overlap",]
  biomart_genes_filt$bias <- biomart_genes_filt$num_cg


  # just entrez genes and number of overlapping CpG
  test_uni_freq <- biomart_genes_filt %>% .[order(.$entrez_id),] %>% group_by(entrez_id) %>% distinct(entrez_id, bias) %>% ungroup() %>% data.frame() # probs slow
  test_uni_freq <- test_uni_freq[test_uni_freq$bias > 0,]
  test_uni_freq

  freq_make_table <- rep(test_uni_freq$entrez_id, test_uni_freq$bias) #this has replaced the loop and sped it up

  freq_make_table <- table(freq_make_table)
  biomart_genes_filt <- biomart_genes_filt[biomart_genes_filt$bias > 0,]
  test_anno_filt <- test_anno_filt[test_anno_filt$ensembl_gene_id %in% biomart_genes_filt$ensembl_gene_id,]
  biomart_genes_filt$is_de <- ifelse(biomart_genes_filt$entrez_id %in% test_anno_filt$entrez_id, 1, 0)
  biomart_genes_filt$is_de <- as.integer(biomart_genes_filt$is_de)
  out <- list(
    #fract.counts = biomart_genes_filt %>%
    # mutate(sigid = entrez_id) %>% distinct(sigid) %>% .[order(.$sigid),] %>% data.frame() %>%
    # mutate(frac = as.double(1)),
    sig.eg = test_anno_filt %>% distinct(entrez_id) %>% .[order(.$entrez_id),],
    universe = biomart_genes_filt %>% distinct(entrez_id) %>%
      .[order(.$entrez_id),] %>% .$entrez_id,
    freq = freq_make_table,
    equiv = as.array(freq_make_table), # column names wrong but maybe not too bad?
    de = biomart_genes_filt[order(biomart_genes_filt$entrez_id),] %>% .$is_de
  )
  out
}

# does all the testing
run_miss_methyl_b <- function(out, method, plot_bias=FALSE, ...) {
  sorted.eg.sig <- out$sig.eg #entrez ids for sig cpgs - genes that have sig cpgs
  eg.universe <- out$universe # entrez ids all genes
  freq_genes <- out$freq # num cpgs per gene
  test.de <- out$de #gene de or not
  #frac <- out$fract.counts
  equiv <- out$equiv
  ## setting up stuff

  # various go filters
  # Check collection is a list with character vectors
  go <- .getGO()
  collection <- go$idList
  #only if it's not a list
  #collection <- list(collection=go$idList)
  collection <- lapply(collection, as.character)
  # Make sure gene set collections don't have any NAs
  collection <- lapply(collection, function(x) x[!is.na(x)])
  # Remove genes that are NOT in the universe from collections
  collection <- lapply(collection, function(x) x[x %in% eg.universe])
  # Remove collections with no genes left after universe filter
  inUniv <- sapply(collection, function(x) length(x) > 0)
  collection <- collection[inUniv]

  # set up testing
  results <- matrix(NA,ncol=4,nrow=length(collection))
  colnames(results) <- c("N", "DE", "P.DE","FDR")
  rownames(results) <- names(collection)
  results[,"N"] <- unlist(lapply(collection,length))
  SigGenesInSet <- rep(NA,length(collection)) # this one is optional

  results[,"DE"] <- unlist(lapply(collection, function(x)
    sum((sorted.eg.sig %in% x))))
  Nuniverse <- length(eg.universe)
  m <- length(sorted.eg.sig)

  # speed up testing by removing categories with 0 de genes (won't be signif)
  # so don't test, and then add them back at the end
  results <- data.frame(results)
  results_null <- results[results$DE == 0,]
  results_null$P.DE <- 1
  results_null$FDR <- 1

  results <- results[results$DE != 0,]
  collection <- collection[rownames(results)]
  results <- as.matrix(results)

  # testing

  if(method != "None") {
    test_pwf <- .estimatePWF(D=test.de,bias=as.vector(freq_genes))
    plot <- .plotBias(D = test.de, bias = as.vector(freq_genes))
  }
  if(method == "Wallenius") {
    results[, c("P.DE")] <-
      t(sapply(1:length(collection), function(i) {
        InSet <- eg.universe %in% collection[[i]] # genes in category
        pw.red <- sum(test_pwf[InSet], na.rm = TRUE) / results[i, "N"] # so i don't know why I get na's - or do i?
        pw.white <- sum(test_pwf[!InSet], na.rm = TRUE) / (Nuniverse - results[i, "N"]) #anyway the na removal is not in original code
        odds <- pw.red / pw.white
        p_val <- BiasedUrn::pWNCHypergeo(results[i,"DE"],
                                         results[i,"N"],
                                         Nuniverse-results[i,"N"],
                                         m,odds,lower.tail=FALSE) +
          BiasedUrn::dWNCHypergeo(results[i,"DE"],
                                  results[i,"N"],
                                  Nuniverse-results[i,"N"],
                                  m,odds)
        #the optional thing
        # Get gene symbols of significant genes
        SigGenesEntrezID <- sorted.eg.sig[sorted.eg.sig %in% collection[[i]]]
        SigGenesSymbol <- suppressMessages(AnnotationDbi::select(org.Hs.eg.db,
                                                                 keys = SigGenesEntrezID,
                                                                 columns = "SYMBOL"))
        SigGenesInSet[i] <- paste(SigGenesSymbol$SYMBOL,collapse=",")
        if(p_val==0){
          message("Pvalue of exactly zero detected. Performing hypergeometric
                test for gene set ", rownames(results)[i])
          p_val <- stats::phyper(q=results[i,"DE"]-0.5,m=m,
                                 n=Nuniverse-m,k=results[i,"N"],
                                 lower.tail=FALSE)
        }
        #  }
        c(p_val)  # return the values as a vector
      })) #chatgpt translated loop to apply
  } else if(method == "None") {
    results[, c("P.DE")] <-
      t(sapply(1:length(collection), function(i) {
        p_val <- stats::phyper(q=results[i,"DE"]-0.5,m=m,
                               n=Nuniverse-m,k=results[i,"N"],
                               lower.tail=FALSE)
        #the optional thing
        # Get gene symbols of significant genes
        SigGenesEntrezID <- sorted.eg.sig[sorted.eg.sig %in% collection[[i]]]
        SigGenesSymbol <- suppressMessages(AnnotationDbi::select(org.Hs.eg.db,
                                                                 keys = SigGenesEntrezID,
                                                                 columns = "SYMBOL"))
        SigGenesInSet[i] <- paste(SigGenesSymbol$SYMBOL,collapse=",")
        c(p_val)
      }))
  } else if(method == "Fishers") {
    results[, c("P.DE")] <-
      t(sapply(1:length(collection), function(i) {
        InSet <- eg.universe %in% collection[[i]]
        pw.red <- sum(test_pwf[InSet], na.rm = TRUE) / results[i, "N"] # so i don't know why I get na's - or do i?
        pw.white <- sum(test_pwf[!InSet], na.rm = TRUE) / (Nuniverse - results[i, "N"]) #anyway the na removal is not in original code
        odds <- pw.red / pw.white
        p_val <- BiasedUrn::pFNCHypergeo(results[i,"DE"],
                                         results[i,"N"],
                                         Nuniverse-results[i,"N"],
                                         m,odds,lower.tail=FALSE) +
          BiasedUrn::dFNCHypergeo(results[i,"DE"],
                                  results[i,"N"],
                                  Nuniverse-results[i,"N"],
                                  m,odds)
        #the optional thing
        # Get gene symbols of significant genes
        SigGenesEntrezID <- sorted.eg.sig[sorted.eg.sig %in% collection[[i]]]
        SigGenesSymbol <- suppressMessages(AnnotationDbi::select(org.Hs.eg.db,
                                                                 keys = SigGenesEntrezID,
                                                                 columns = "SYMBOL"))
        SigGenesInSet[i] <- paste(SigGenesSymbol$SYMBOL,collapse=",")
        if(p_val==0){
          message("Pvalue of exactly zero detected. Performing hypergeometric
                test for gene set ", rownames(results)[i])
          p_val <- stats::phyper(q=results[i,"DE"]-0.5,m=m,
                                 n=Nuniverse-m,k=results[i,"N"],
                                 lower.tail=FALSE)
        }
        c(p_val)  # return the values as a vector
      }))
  }
## Introduction

```{r}

```
