==> analysis/04_compareStatTests.Rmd <==
---
title: "04_compareStatTests"
author: "Caitlin Page"
date: "2025-02-10"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: inline
---

## Packages

```{r}
library(BiasedUrn)
library(missMethyl)
library(goseq)
library(GO.db)

library(rbenchmark)

library(dplyr)
library(ggVennDiagram)
library(ggplot2)
```

## Data
* need inputs for GST
```{r}
DMPs <- readRDS("../output/array_dmps.rds")
```


## Microarray

```{r}
topCpGs <- DMPs[1:10000,]
sigCpGs <- rownames(topCpGs)
```
```{r}
mapped_entrez <- getMappedEntrezIDs(sig.cpg=sigCpGs, all.cpg=rownames(DMPs))
```

```{r}
gst <- gometh(sig.cpg=sigCpGs, all.cpg=rownames(DMPs), collection="GO",
              plot.bias=TRUE)
```

```{r}
sig_terms_array <- topGSA(gst, n = 1000) %>% filter(FDR < 0.05)
sig_terms_array
sig_terms_array <- readRDS("../output/mm_array_res.rds")
```


# Goseq version to compare approaches
- except I need the genes and that's under the hood
```{r}
mapped_entrez$freq %>% data.frame()
```
```{r}
length(mapped_entrez$sig.eg)
length(mapped_entrez$de)
length(mapped_entrez$universe)
```

```{r}
gene.vector <- mapped_entrez$de
names(gene.vector) <- mapped_entrez$universe

pwf <- goseq::nullp(gene.vector, "hg19", "Entrez", bias.data = data.frame(mapped_entrez$freq)$Freq)
pwf
GO.wall <- goseq::goseq(pwf, "hg19", "Entrez")
GO.wall <- GO.wall %>% dplyr::mutate(FDR = stats::p.adjust(.data$over_represented_pvalue, method = "BH"))
GO.wall
```
```{r}
genes_symbol <- AnnotationDbi::select(org.Hs.eg.db, 
                                keys=AnnotationDbi::keys(org.Hs.eg.db), 
                                columns=c("ENTREZID","SYMBOL"), 
                                keytype="ENTREZID")

```

```{r}
universe <- data.frame(mapped_entrez$freq)
universe$de <- mapped_entrez$de
universe$eg.all <- as.character(universe$eg.all)
universe$symbol <- genes_symbol[match(universe$eg.all, genes_symbol$ENTREZID), "SYMBOL"]
universe
```


```{r}
gene.vector <- universe$de
names(gene.vector) <- universe$symbol
pwf <- goseq::nullp(gene.vector, "hg19", "geneSymbol", bias.data = universe$Freq)
pwf
GO.wall <- goseq::goseq(pwf, "hg19", "geneSymbol")
GO.wall <- GO.wall %>% dplyr::mutate(FDR = stats::p.adjust(.data$over_represented_pvalue, method = "BH"))
GO.wall
```
```{r}
GO.samp <- goseq(pwf,"hg19","geneSymbol",method="Sampling",repcnt=1000)
```
```{r}
GO.samp <- GO.samp %>% dplyr::mutate(FDR = stats::p.adjust(.data$over_represented_pvalue, method = "BH"))
```

```{r}
GO.fish <- goseq_mod(pwf,"hg19","geneSymbol",method="Fishers")
GO.fish <- GO.fish %>% dplyr::mutate(FDR = stats::p.adjust(.data$over_represented_pvalue, method = "BH"))
```
```{r}
system.time(goseq(pwf, "hg19", "geneSymbol"))
system.time(goseq(pwf,"hg19","geneSymbol",method="Sampling",repcnt=1000))
system.time(goseq_mod(pwf,"hg19","geneSymbol",method="Fishers"))
```
```{r}
benchmark(wall = goseq(pwf, "hg19", "geneSymbol"), fish = goseq_mod(pwf,"hg19","geneSymbol",method="Fishers"), replications = 50)
```
* interesting - basically no difference in timing

## compare results - fishers and wallenius
* can I do a correlation?
* because that's basically just comparing how similar 2 lists are
* if wants numbers - can just give diff terms diff numbers
* ok its more an association thing but it could still work

```{r}
GO.wall <- GO.wall %>% .[order(.$FDR),] %>% mutate(wall_rank = 1:n())
GO.samp <- GO.samp %>% .[order(.$FDR),] %>% mutate(samp_rank = 1:n())
GO.fish <- GO.fish %>% .[order(.$FDR),] %>% mutate(fish_rank = 1:n())
```

```{r}
GO.wall %>% mutate(samp = GO.samp[match(GO.wall$category, GO.samp$category), "samp_rank"],
                   fish = GO.fish[match(GO.wall$category, GO.fish$category), "fish_rank"]) %>%
  ggplot(aes(x = wall_rank, y = samp)) +
  geom_point()

GO.wall %>% mutate(samp = GO.samp[match(GO.wall$category, GO.samp$category), "samp_rank"],
                   fish = GO.fish[match(GO.wall$category, GO.fish$category), "fish_rank"]) %>%
  ggplot(aes(x = wall_rank, y = fish)) +
  geom_point()

GO.wall %>% mutate(samp = GO.samp[match(GO.wall$category, GO.samp$category), "samp_rank"],
                   fish = GO.fish[match(GO.wall$category, GO.fish$category), "fish_rank"]) %>%
  ggplot(aes(x = samp, y = fish)) +
  geom_point()
```
```{r}
ggVennDiagram(list(wall = filter(GO.wall, FDR < 0.05)$category, 
              samp = filter(GO.samp, FDR < 0.05)$category,
              fish = filter(GO.fish, FDR < 0.05)$category))
```
```{r}
ggVennDiagram(list(wall = filter(MM_wall_filt, FDR < 0.05)$GOID, 
              none = filter(MM_none_filt, FDR < 0.05)$GOID,
              fish = filter(MM_fish_filt, FDR < 0.05)$GOID))
```

```{r}
MM_wall_filt
```

```{r}
GO.wall 
GO.samp
GO.fish
```
* GO.samp - the results after adjustment are not significant
* GOseq PDF: "However,
the accuracy of the sampling method is limited by the number of samples generated, repcnt, such
that very low p-values will not be correctly calculated. Significantly enriched GO terms may then
be missed after correcting for multiple testing."
* this may be why nothing is significant
* but wasn't the sampling supposed to be the best version - and Wallenius an approximation?
```{r}
plot(log10(GO.wall[,2]), log10(GO.samp[match(GO.wall[,1],GO.samp[,1]),2]),
xlab="log10(Wallenius p-values)",ylab="log10(Sampling p-values)",
xlim=c(-3,0))
abline(0,1,col=3,lty=2)
```

## WGBS
```{r}
wgbs_counts <- readRDS("../output/wgbs_counts.rds")
bsseq_dmrs <- readRDS("../output/bsseq_dmrs.rds")
biomart_genes <- readRDS("../output/genes_biomart.rds")
```
```{r anno}
source("../code/function_binBiasPlot.R")
test_anno <- annoGeneDmr(wgbs_counts, bsseq_dmrs, gene_source = "biomaRt", genes = biomart_genes)
```
```{r}
test_anno_filt <- test_anno %>% filter(areaStat > 100)
```


```{r}
test_bias <- biomart_genes %>% mutate(width = end - start + 1, has_dmr = ifelse(ensembl_gene_id %in% filter(test_anno, abs(min_dist) == 0)$ensembl_gene_id, TRUE, FALSE)) %>%
    group_by(num_cg) %>%
    mutate(num_bins_same_cg = dplyr::n()) %>%
    .[order(.$num_cg),] %>%
    ungroup() %>%
    mutate(bin_group = rep(1:ceiling(nrow(biomart_genes)/100), each = 100)[1:nrow(biomart_genes)]) %>%
    group_by(bin_group) %>%
    mutate(num_cg_bin_group = mean(num_cg)) %>%
    ungroup() %>% mutate(has_dmr = ifelse(has_dmr == FALSE, 0, 1))
test_bias
test_bias_filt <- biomart_genes %>% mutate(width = end - start + 1, has_dmr = ifelse(ensembl_gene_id %in% filter(test_anno_filt, abs(min_dist) == 0)$ensembl_gene_id, TRUE, FALSE)) %>%
    group_by(num_cg) %>%
    mutate(num_bins_same_cg = dplyr::n()) %>%
    .[order(.$num_cg),] %>%
    ungroup() %>%
    mutate(bin_group = rep(1:ceiling(nrow(biomart_genes)/100), each = 100)[1:nrow(biomart_genes)]) %>%
    group_by(bin_group) %>%
    mutate(num_cg_bin_group = mean(num_cg)) %>%
    ungroup() %>% mutate(has_dmr = ifelse(has_dmr == FALSE, 0, 1))
test_bias_filt
```

# goseq
```{r}
gene.vector <- test_bias_filt$has_dmr
names(gene.vector) <- test_bias_filt$ensembl_gene_id

pwf_seq <- goseq::nullp(gene.vector, "hg19", "ensGene", bias.data = test_bias_filt$num_cg_bin_group)
pwf_seq
GO.wall_seq <- goseq::goseq(pwf_seq, "hg19", "ensGene")
GO.wall_seq <- GO.wall_seq %>% dplyr::mutate(FDR = stats::p.adjust(.data$over_represented_pvalue, method = "BH"))

GO.samp_seq <- goseq(pwf_seq,"hg19","ensGene",method="Sampling",repcnt=1000)
GO.samp_seq <- GO.samp_seq %>% dplyr::mutate(FDR = stats::p.adjust(.data$over_represented_pvalue, method = "BH"))

GO.fish_seq <- goseq_mod(pwf_seq,"hg19","ensGene",method="Fishers")
GO.fish_seq <- GO.fish_seq %>% dplyr::mutate(FDR = stats::p.adjust(.data$over_represented_pvalue, method = "BH"))

GO.wall_seq <- GO.wall_seq %>% .[order(.$FDR),] %>% mutate(wall_rank = 1:dplyr::n())
GO.samp_seq <- GO.samp_seq %>% .[order(.$FDR),] %>% mutate(samp_rank = 1:dplyr::n())
GO.fish_seq <- GO.fish_seq %>% .[order(.$FDR),] %>% mutate(fish_rank = 1:dplyr::n())
```

```{r}
system.time(goseq(pwf_seq, "hg19", "ensGene"))
system.time(goseq(pwf_seq,"hg19","ensGene",method="Sampling",repcnt=1000))
system.time(goseq_mod(pwf_seq,"hg19","ensGene",method="Fishers"))
```

```{r}
benchmark(wall = goseq(pwf_seq, "hg19", "ensGene"), fish = goseq_mod(pwf_seq,"hg19","ensGene",method="Fishers"), replications = 50)
```


# missMethyl
```{r}
test_out <- readRDS("../output/mm_fn_out_num.rds")
test_out_filt <- run_miss_methyl_a(biomart_genes, filter(test_anno, abs(areaStat) > 100))
```

```{r}
MM_fish_filt <- run_miss_methyl_b(test_out_filt, method = "Fishers")
MM_wall_filt <- run_miss_methyl_b(test_out_filt, method = "Fishers")
MM_none_filt <- run_miss_methyl_b(test_out_filt, method = "None")
```

```{r}
system.time(run_miss_methyl_b(test_out, method = "Wallenius"))
system.time(run_miss_methyl_b(test_out, method = "Fishers"))
```
* ok this implies no difference still
```{r}
system.time(run_miss_methyl_b(test_out_filt, method = "Wallenius"))
system.time(run_miss_methyl_b(test_out_filt, method = "Fishers"))
```
* damn now it's not even faster?
* trying to skip testing for 0 de for wallenius only
```{r}
system.time(run_miss_methyl_b(test_out_filt, method = "Wallenius"))
system.time(run_miss_methyl_b(test_out_filt, method = "Fishers"))
```

* damn that made it longer than the previous - and worse it made the system time longer
* ok but the other one took longer this time as well

* but also thought of better way - just remove them before the testing
```{r}
system.time(run_miss_methyl_b(test_out_filt, method = "Wallenius"))
system.time(run_miss_methyl_b(test_out_filt, method = "Fishers"))
```
* haha that is faster system wise
* still slow overall
* still don't get why it's so much slower than goseq
* and now fishers is slower than wall again
* why is it so much slower than goseq??
```{r}
GO.wall
```

```{r}
run_miss_methyl_b(test_out_filt, method = "Wallenius")
```

```{r}
test_out$freq %>% data.frame()
test_out$de %>% data.frame()
```
```{r}
universe
```
```{r}
GO.fish %>% filter(numDEInCat == 0)
```

```{r}
MM_fish %>% filter(DE == 0) %>% .[order(.$N),]
```



## Code
# seq code - missmethyl

* if Fishers is independent, I should be able to not do it as a loop or whatever
* so I can have a vector for x - which is the num of red balls sampled
* but we put in the de? - right the number of de genes
* but this wouldn't change? because it's our list? - or its got something to do with something
* maybe it still needs to be a loop because it's still an individual test for each category and we need a p val for each
* and the speed should be that something else is happening under the hood of the test which might be faster
* or maybe I can // it?
*I'm not figuring out what goes under the hood of the tests - found the code - I think it's written in C++
* I do not read C++
* can say the Wallenius file is longer than the Fishers
* ok so reviewed the what goes in and out and that all makes sense
* guessing minimal speed difference in methods is because of the c++ efficiency

# what to do for speed?
* prefilter out go categories?? - so not test them if low?
* can't tell if that's what is currently happening in goseq - all 1 for p, mm still has a p


  if(prior.prob){
    for(i in 1:length(collection)){
      InSet <- eg.universe %in% collection[[i]]
      pw.red <- sum(pwf[InSet])/results[i,"N"]
      pw.white <- sum(pwf[!InSet])/(Nuniverse-results[i,"N"])
      odds <- pw.red/pw.white
      results[i,"P.DE"] <- BiasedUrn::pWNCHypergeo(results[i,"DE"],
                                                   results[i,"N"],
                                                   Nuniverse-results[i,"N"],
                                                   m,odds,lower.tail=FALSE) + 
          BiasedUrn::dWNCHypergeo(results[i,"DE"],
                                  results[i,"N"],
                                  Nuniverse-results[i,"N"],
                                  m,odds)
  }
  }
```{r}
system.time(go_stuff(test_out_filt))
```
```{r}
rbind(MM_fish[MM_fish$DE != 0,], MM_fish[MM_fish$DE == 0,])
```
```{r}
try_res <- try_speed(test_out_filt)
```
```{r}
class(try_res)
```

```{r}
try_res %>% data.frame() %>% .[.$DE != 0,] %>% as.matrix() %>% class()
data.frame(NA,ncol=4,nrow=10)
```
```{r}
try_collect <- try_speed(test_out_filt)
```

```{r}
go_stuff <- function(out) {
  sorted.eg.sig <- out$sig.eg #entrez ids for sig cpgs - genes that have sig cpgs
  eg.universe <- out$universe # entrez ids all cpgs(?) - but it's just all genes
  freq_genes <- out$freq
  test.de <- out$de #res for each gene
  #frac <- out$fract.counts
  equiv <- out$equiv
  ## setting up stuff
  # Check collection is a list with character vectors
  go <- .getGO()
  collection <- go$idList  
  #only if it's not a list
  #collection <- list(collection=go$idList)
  collection <- lapply(collection, as.character)
  # Make sure gene set collections don't have any NAs
  collection <- lapply(collection, function(x) x[!is.na(x)])
  # Remove genes that are NOT in the universe from collections
  collection <- lapply(collection, function(x) x[x %in% eg.universe])
  # Remove collections with no genes left after universe filter
  inUniv <- sapply(collection, function(x) length(x) > 0)
  collection <- collection[inUniv]
}

try_speed <- function(out) {
  sorted.eg.sig <- out$sig.eg #entrez ids for sig cpgs - genes that have sig cpgs
  eg.universe <- out$universe # entrez ids all cpgs(?) - but it's just all genes
  freq_genes <- out$freq
  test.de <- out$de #res for each gene
  #frac <- out$fract.counts
  equiv <- out$equiv
  ## setting up stuff
  # Check collection is a list with character vectors
  go <- .getGO()
  collection <- go$idList  
  #only if it's not a list
  #collection <- list(collection=go$idList)
  collection <- lapply(collection, as.character)
  # Make sure gene set collections don't have any NAs
  collection <- lapply(collection, function(x) x[!is.na(x)])
  # Remove genes that are NOT in the universe from collections
  collection <- lapply(collection, function(x) x[x %in% eg.universe])
  # Remove collections with no genes left after universe filter
  inUniv <- sapply(collection, function(x) length(x) > 0)
  collection <- collection[inUniv]
  
  
  results <- matrix(NA,ncol=4,nrow=length(collection))
  colnames(results) <- c("N", "DE", "P.DE","FDR")
  rownames(results) <- names(collection)
  results[,"N"] <- unlist(lapply(collection,length))
  SigGenesInSet <- rep(NA,length(collection)) # this one is optional
  
  results[,"DE"] <- unlist(lapply(collection, function(x) 
          sum((sorted.eg.sig %in% x))))
  Nuniverse <- length(eg.universe)
  m <- length(sorted.eg.sig)
  collection
}
```
```{r}
data.frame(try_res) %>% .[.$DE != 0,] %>% as.matrix() %>% nrow()
```
```{r}
length(try_collect)
length(try_collect[data.frame(try_res) %>% .[.$DE != 0,] %>% rownames()])
```

```{r}
.getGO <- function(){
  if(!requireNamespace("org.Hs.eg.db", quietly = TRUE))
    stop("org.Hs.eg.db package required but not installed.")
  egGO2ALLEGS <- utils::getFromNamespace("org.Hs.egGO2ALLEGS", "org.Hs.eg.db")
  GeneID.PathID <- AnnotationDbi::toTable(egGO2ALLEGS)[,c("gene_id", "go_id", "Ontology")]
  d <- !duplicated(GeneID.PathID[, c("gene_id", "go_id")])
  GeneID.PathID <- GeneID.PathID[d, ]
  GOID.TERM <- suppressMessages(AnnotationDbi::select(GO.db::GO.db, 
                                                      keys=unique(GeneID.PathID$go_id), 
                                                      columns=c("GOID","ONTOLOGY","TERM"), 
                                                      keytype="GOID"))
  go <- tapply(GeneID.PathID$gene_id, GeneID.PathID$go_id, list)
    
  list(idList=go, idTable=GOID.TERM)
}

.estimatePWF <- function(D,bias)
  # An alternative to goseq function nullp, which is transformation invariant
  # Belinda Phipson and Gordon Smyth
  # 6 March 2015
{
  prior.prob <- bias
  o <- order(bias)
  prior.prob[o] <- limma::tricubeMovingAverage(D[o],span=0.5)
  prior.prob
}

run_miss_methyl_b <- function(out, method, plot_bias=FALSE, ...) {
  sorted.eg.sig <- out$sig.eg #entrez ids for sig cpgs - genes that have sig cpgs
  eg.universe <- out$universe # entrez ids all cpgs(?) - but it's just all genes
  freq_genes <- out$freq
  test.de <- out$de #res for each gene
  #frac <- out$fract.counts
  equiv <- out$equiv
  ## setting up stuff
  # Check collection is a list with character vectors
  go <- .getGO()
  collection <- go$idList  
  #only if it's not a list
  #collection <- list(collection=go$idList)
  collection <- lapply(collection, as.character)
  # Make sure gene set collections don't have any NAs
  collection <- lapply(collection, function(x) x[!is.na(x)])
  # Remove genes that are NOT in the universe from collections
  collection <- lapply(collection, function(x) x[x %in% eg.universe])
  # Remove collections with no genes left after universe filter
  inUniv <- sapply(collection, function(x) length(x) > 0)
  collection <- collection[inUniv]
  
  
  results <- matrix(NA,ncol=4,nrow=length(collection))
  colnames(results) <- c("N", "DE", "P.DE","FDR")
  rownames(results) <- names(collection)
  results[,"N"] <- unlist(lapply(collection,length))
  SigGenesInSet <- rep(NA,length(collection)) # this one is optional
  
  results[,"DE"] <- unlist(lapply(collection, function(x) 
          sum((sorted.eg.sig %in% x))))
  Nuniverse <- length(eg.universe)
  m <- length(sorted.eg.sig)
  
  results <- data.frame(results)
  results_null <- results[results$DE == 0,]
  results_null$P.DE <- 1
  results_null$FDR <- 1
  
  results <- results[results$DE != 0,]
  collection <- collection[rownames(results)]
  results <- as.matrix(results)
  
  if(method != "None") {
  test_pwf <- .estimatePWF(D=test.de,bias=as.vector(freq_genes))
  #  if(plot_bias == TRUE) {
   #   plotBiasGrouped(genes = genes, anno_genes = , ...)
  #  }
  # the testing
  }
  if(method == "Wallenius") {
  results[, c("P.DE")] <- 
  t(sapply(1:length(collection), function(i) {
    # try to skip the 0 des to speed it up
 #   if(results[i,"DE"] == 0) {
 #     p_val <- 1
  #  } else {
    InSet <- eg.universe %in% collection[[i]]
    pw.red <- sum(test_pwf[InSet], na.rm = TRUE) / results[i, "N"] # so i don't know why I get na's - or do i?
    pw.white <- sum(test_pwf[!InSet], na.rm = TRUE) / (Nuniverse - results[i, "N"]) #anyway the na removal is not in original code
    odds <- pw.red / pw.white
    p_val <- BiasedUrn::pWNCHypergeo(results[i,"DE"],
                                                   results[i,"N"],
                                                   Nuniverse-results[i,"N"],
                                                   m,odds,lower.tail=FALSE) + 
          BiasedUrn::dWNCHypergeo(results[i,"DE"],
                                  results[i,"N"],
                                  Nuniverse-results[i,"N"],
                                  m,odds)
    #the optional thing
      # Get gene symbols of significant genes
    SigGenesEntrezID <- sorted.eg.sig[sorted.eg.sig %in% collection[[i]]]
    SigGenesSymbol <- suppressMessages(AnnotationDbi::select(org.Hs.eg.db, 
                                             keys = SigGenesEntrezID,
                                             columns = "SYMBOL"))
    SigGenesInSet[i] <- paste(SigGenesSymbol$SYMBOL,collapse=",")
    if(p_val==0){
        message("Pvalue of exactly zero detected. Performing hypergeometric 
                test for gene set ", rownames(results)[i])
        p_val <- stats::phyper(q=results[i,"DE"]-0.5,m=m,
                                           n=Nuniverse-m,k=results[i,"N"],
                                           lower.tail=FALSE)
    }
  #  }
    c(p_val)  # return the values as a vector
    })) #chatgpt translated loop to apply
  } else if(method == "None") {
    results[, c("P.DE")] <- 
    t(sapply(1:length(collection), function(i) {
      p_val <- stats::phyper(q=results[i,"DE"]-0.5,m=m,
                                           n=Nuniverse-m,k=results[i,"N"],
                                           lower.tail=FALSE)
      #the optional thing
      # Get gene symbols of significant genes
    SigGenesEntrezID <- sorted.eg.sig[sorted.eg.sig %in% collection[[i]]]
    SigGenesSymbol <- suppressMessages(AnnotationDbi::select(org.Hs.eg.db, 
                                             keys = SigGenesEntrezID,
                                             columns = "SYMBOL"))
    SigGenesInSet[i] <- paste(SigGenesSymbol$SYMBOL,collapse=",")
      c(p_val)
    }))
  } else if(method == "Fishers") {
    results[, c("P.DE")] <- 
  t(sapply(1:length(collection), function(i) {
    InSet <- eg.universe %in% collection[[i]]
    pw.red <- sum(test_pwf[InSet], na.rm = TRUE) / results[i, "N"] # so i don't know why I get na's - or do i?
    pw.white <- sum(test_pwf[!InSet], na.rm = TRUE) / (Nuniverse - results[i, "N"]) #anyway the na removal is not in original code
    odds <- pw.red / pw.white
    p_val <- BiasedUrn::pFNCHypergeo(results[i,"DE"],
                                                   results[i,"N"],
                                                   Nuniverse-results[i,"N"],
                                                   m,odds,lower.tail=FALSE) + 
          BiasedUrn::dFNCHypergeo(results[i,"DE"],
                                  results[i,"N"],
                                  Nuniverse-results[i,"N"],
                                  m,odds)
    #the optional thing
      # Get gene symbols of significant genes
    SigGenesEntrezID <- sorted.eg.sig[sorted.eg.sig %in% collection[[i]]]
    SigGenesSymbol <- suppressMessages(AnnotationDbi::select(org.Hs.eg.db, 
                                             keys = SigGenesEntrezID,
                                             columns = "SYMBOL"))
    SigGenesInSet[i] <- paste(SigGenesSymbol$SYMBOL,collapse=",")
    if(p_val==0){
        message("Pvalue of exactly zero detected. Performing hypergeometric 
                test for gene set ", rownames(results)[i])
        p_val <- stats::phyper(q=results[i,"DE"]-0.5,m=m,
                                           n=Nuniverse-m,k=results[i,"N"],
                                           lower.tail=FALSE)
    }
    c(p_val)  # return the values as a vector
    }))
  }
  
  
  results <- results %>% data.frame()
  results[,"FDR"] <- stats::p.adjust(results[,"P.DE"],method="BH")
  results[,"DE"] <- floor(results[,"DE"])
  results <- rbind(results, results_null)
  results <- data.frame(results, SigGenesInSet)
  
  results <- merge(go$idTable,results,by.x="GOID",by.y="row.names")
  rownames(results) <- results$GOID
  results <- results %>% .[order(.$FDR),]
  results
  
}

```





```{r}
columns(org.Hs.eg.db)
```
```{r}
keys(org.Hs.eg.db)
```
