==> analysis/04a_compareGSTResults.Rmd <==
---
title: "04a_compareGSTResults"
author: "Caitlin Page"
date: "2024-11-29"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: inline
---

## Packages
```{r}
library(plyranges)
library(dplyr)
library(ggplot2)
library(ggVennDiagram)
library(reshape2)
```

## Description


## Files
```{r}
go_entrez <- readRDS("../output/go_entrez.rds")
```

```{r}
GO.wall <- readRDS("../output/goseq_res.rds")
GO.wall.null <- readRDS("../output/goseq_null_res.rds")

mm_hack <- readRDS("../output/mm_hack_res.rds")
mm_hack_null <- readRDS("../output/mm_hack_null_res.rds")

sig_terms_array <- readRDS("../output/mm_array_res.rds")
sig_terms_array_biased <- readRDS("../output/mm_array_null.rds")
```


## Introduction
```{r}
go_entrez
```
# 
```{r}
go_entrez[,1:6] %>% group_by(go) %>% mutate(mean_cg = mean(num_cg, na.rm=TRUE)) %>% distinct(go, array_cat_signif, array_cat_signif_bias, mean_cg) %>% filter(array_cat_signif_bias == TRUE) %>% 
  mutate(Method = ifelse(array_cat_signif == FALSE, "Biased (Hypergeometric)", "Bias corrected (Wallenius)")) %>%
  ggplot(aes(x = Method, y = mean_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(y = "Average number of CpGs per ontology category")
```
```{r}
go_entrez[,1:6] %>% group_by(go) %>% mutate(mean_cg = mean(num_cg, na.rm=TRUE)) %>% distinct(go, array_cat_signif, array_cat_signif_bias, mean_cg) %>%
  
```

# looking at significant categories - don't care if signif in both - just if it's signif in unbiased or not - and the mean cg per category
```{r}
go_entrez[,1:6] %>% group_by(go) %>% mutate(mean_cg = mean(num_cg, na.rm=TRUE)) %>% distinct(go, array_cat_signif, array_cat_signif_bias, mean_cg) %>% filter(array_cat_signif_bias == TRUE | array_cat_signif == TRUE) %>%
  mutate(Method = ifelse(array_cat_signif == FALSE, "Biased (Hypergeometric)", "Bias corrected (Wallenius)")) %>%
  ggplot(aes(x = Method, y = mean_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(y = "Average number of CpGs per ontology category", title = "MM Array")

go_entrez %>% group_by(go) %>% mutate(mean_cg = mean(num_cg, na.rm=TRUE)) %>% distinct(go, seq_cat_signif, seq_cat_signif_bias, mean_cg) %>% filter(seq_cat_signif_bias == TRUE | seq_cat_signif == TRUE) %>%
  mutate(Method = ifelse(seq_cat_signif == FALSE, "Biased (Hypergeometric)", "Bias corrected (Wallenius)")) %>%
  ggplot(aes(x = Method, y = mean_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(y = "Average number of CpGs per ontology category", title = "Seq Goseq")

go_entrez %>% group_by(go) %>% mutate(mean_cg = mean(num_cg, na.rm=TRUE)) %>% distinct(go, mm_seq, mm_seq_null, mean_cg) %>% filter(mm_seq == TRUE | mm_seq_null == TRUE) %>%
  mutate(Method = ifelse(mm_seq == FALSE, "Biased (Hypergeometric)", "Bias corrected (Wallenius)")) %>%
  ggplot(aes(x = Method, y = mean_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(y = "Average number of CpGs per ontology category", title = "Hacked MM")

go_entrez %>% group_by(go) %>% mutate(mean_cg = mean(num_cg_bin, na.rm=TRUE)) %>% distinct(go, mm_fn_bin, mm_fn_bin_null, mean_cg) %>% filter(mm_fn_bin == TRUE | mm_fn_bin_null == TRUE) %>%
  mutate(Method = ifelse(mm_fn_bin == FALSE, "Biased (Hypergeometric)", "Bias corrected (Wallenius)")) %>%
  ggplot(aes(x = Method, y = mean_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(y = "Average number of CpGs per ontology category", title = "Fn MM cg_bin")

go_entrez %>% group_by(go) %>% mutate(mean_cg = mean(num_cg, na.rm=TRUE)) %>% distinct(go, mm_fn_num, mm_fn_num_null, mean_cg) %>% filter(mm_fn_num == TRUE | mm_fn_num_null == TRUE) %>%
  mutate(Method = ifelse(mm_fn_num == FALSE, "Biased (Hypergeometric)", "Bias corrected (Wallenius)")) %>%
  ggplot(aes(x = Method, y = mean_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(y = "Average number of CpGs per ontology category", title = "Fn MM num_cg")
```
```{r}
go_entrez %>% group_by(go) %>% mutate(mean_cg = mean(num_cg, na.rm=TRUE)) %>% distinct(go, mm_fn_num, mm_fn_num_null, mean_cg) %>% filter(mm_fn_num == TRUE | mm_fn_num_null == TRUE) %>%
  mutate(Method = ifelse(mm_fn_num == FALSE, "Biased (Hypergeometric)", "Bias corrected (Wallenius)")) %>%
  ggplot(aes(x = Method, y = mean_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(y = "Average number of CpGs per ontology category", title = "Fn MM num_cg")
```
```{r}
go_entrez %>% group_by(go) %>% mutate(mean_cg = mean(num_cg, na.rm=TRUE)) %>% distinct(go, mm_fn_num, mm_fn_num_null, mean_cg) %>% filter(mm_fn_num == TRUE | mm_fn_num_null == TRUE) %>%
  mutate(Method = ifelse(mm_fn_num == FALSE, "Biased (Hypergeometric)", "Bias corrected (Wallenius)"))
```
```{r}
go_entrez %>% group_by(go) %>% mutate(mean_cg = mean(num_cg, na.rm=TRUE)) %>% distinct(go, mm_fn_num_filt_dmr, mm_fn_num_filt_dmr_null, mean_cg) %>% filter(mm_fn_num_filt_dmr == TRUE | mm_fn_num_filt_dmr_null == TRUE) %>%
  mutate(Method = ifelse(mm_fn_num_filt_dmr == FALSE, "Biased (Hypergeometric)", "Bias corrected (Wallenius)")) %>%
  ggplot(aes(x = Method, y = mean_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(y = "Average number of CpGs per ontology category", title = "Fn MM num_cg filt dmr")
```
- still not as good as array
```{r}
go_entrez
```

```{r}
go_entrez %>% group_by(go) %>% mutate(mean_cg = mean(num_cg, na.rm=TRUE)) %>% distinct(go, mm_fn_num_filt_dmr, mm_fn_num_filt_dmr_null, mean_cg) %>%
  mutate(Method = case_when(mm_fn_num_filt_dmr == FALSE & mm_fn_num_filt_dmr_null == FALSE ~ "Neither",
                            mm_fn_num_filt_dmr == TRUE & mm_fn_num_filt_dmr_null == FALSE ~ "Corrected",
                            mm_fn_num_filt_dmr == FALSE & mm_fn_num_filt_dmr_null == TRUE ~ "Biased",
                            TRUE ~ "Both")) %>%
  ggplot(aes(x = Method, y = mean_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(y = "Average number of CpGs per ontology category", title = "Fn MM num_cg filt dmr")
```
- ooh maybe like this? - try for array
```{r}
go_entrez %>% group_by(go) %>% mutate(mean_cg = mean(num_cg, na.rm=TRUE)) %>% distinct(go, mm_fn_num_filt_dmr, mm_fn_num_filt_dmr_null, mean_cg) %>%
  mutate(Method = case_when(mm_fn_num_filt_dmr == FALSE & mm_fn_num_filt_dmr_null == FALSE ~ "Neither",
                            mm_fn_num_filt_dmr == TRUE & mm_fn_num_filt_dmr_null == FALSE ~ "Corrected",
                            mm_fn_num_filt_dmr == FALSE & mm_fn_num_filt_dmr_null == TRUE ~ "Biased",
                            TRUE ~ "Both")) %>%
  ggplot(aes(x = Method, y = log2(mean_cg + 1), fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(y = "Average number of CpGs per ontology category", title = "Fn MM num_cg filt dmr")
```

```{r}
go_entrez %>% group_by(go) %>% mutate(mean_cg = mean(num_cg, na.rm=TRUE)) %>% distinct(go, array_cat_signif, array_cat_signif_bias, mean_cg) %>%
  mutate(Method = case_when(array_cat_signif == FALSE & array_cat_signif_bias == FALSE ~ "Neither",
                            array_cat_signif == TRUE & array_cat_signif_bias == FALSE ~ "Corrected",
                            array_cat_signif == FALSE & array_cat_signif_bias == TRUE ~ "Biased",
                            TRUE ~ "Both")) %>%
  ggplot(aes(x = Method, y = log2(mean_cg + 1), fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(y = "Average number of CpGs per ontology category", title = "Array")
```
```{r}
go_entrez %>% group_by(go) %>% mutate(mean_cg = mean(num_cg, na.rm=TRUE)) %>% distinct(go, array_cat_signif, array_cat_signif_bias, mean_cg) %>%
  mutate(Method = case_when(array_cat_signif == FALSE & array_cat_signif_bias == FALSE ~ "Neither",
                            array_cat_signif == TRUE & array_cat_signif_bias == FALSE ~ "Corrected",
                            array_cat_signif == FALSE & array_cat_signif_bias == TRUE ~ "Biased",
                            TRUE ~ "Both")) %>%
  ggplot(aes(x = Method, y = mean_cg, fill = Method)) +
  geom_boxplot() +
  theme_bw() +
  labs(y = "Average number of CpGs per ontology category", title = "Array")
```
```{r}
go_entrez %>% group_by(go) %>% mutate(mean_cg = mean(num_cg, na.rm=TRUE)) %>% distinct(go, array_cat_signif, array_cat_signif_bias, mean_cg) %>%
  mutate(Method = case_when(array_cat_signif == FALSE & array_cat_signif_bias == FALSE ~ "Neither",
                            array_cat_signif == TRUE & array_cat_signif_bias == FALSE ~ "Corrected",
                            array_cat_signif == FALSE & array_cat_signif_bias == TRUE ~ "Biased",
                            TRUE ~ "Both")) %>% filter(Method == "Corrected")
```
```{r}
go_entrez[,1:6] %>% filter(array_cat_signif == TRUE) %>% distinct(go, array_cat_signif, array_cat_signif_bias)
```

```{r}
bsseq_dmrs[66000:67000,]
```
```{r}
bsseq_dmrs 
```
```{r}
bsseq_dmrs[20000:21000,] 
```

```{r}
tail(bsseq_dmrs)
```
```{r}
bsseq_dmrs %>%
  ggplot(aes(x = abs(areaStat))) +
  geom_density()
```
```{r}
bsseq_dmrs %>% mutate(less_250 = ifelse(abs(areaStat) <= 100, TRUE, FALSE)) %>% group_by(less_250) %>% summarise(n())
```
```{r}
bsseq_dmrs %>% filter(abs(areaStat) > 100) %>%
  ggplot(aes(x = abs(areaStat))) +
  geom_density()
```


```{r}
go_entrez
test_b_num
test_b_num_null
```



# change in rank plots
```{r}
go_entrez
```
```{r}
go_entrez <- compare
```

```{r}
go_entrez2 <- compare %>% group_by(go) %>% mutate(mean_cg = mean(num_cg, na.rm=TRUE), mean_cg_bin = mean(num_cg_bin, na.rm=TRUE)) %>% ungroup() %>% distinct(go, mean_cg, mean_cg_bin) %>% data.frame()
```
```{r}
go_entrez
```
```{r}
go_entrez2
```


```{r}
sig_terms_array$rank <- 1:nrow(sig_terms_array)
sig_terms_array_biased$rank <- 1:nrow(sig_terms_array_biased)
ggVennDiagram(list(yes = rownames(sig_terms_array), no = rownames(sig_terms_array_biased)))
```
- oh if I didn't filter to significant categories i could have all the ranks here - damn i need to re-run things again
```{r}
sig_terms_array_biased %>% mutate(unbiased_rank = sig_terms_array[match(rownames(.), rownames(sig_terms_array)), "rank"], mean_cg = go_entrez2[match(rownames(.), go_entrez2$go), "mean_cg"]) %>% mutate(rank_change = unbiased_rank - rank) %>%
  ggplot(aes(x = mean_cg, y = rank_change)) +
  geom_point()

sig_terms_array_biased %>% mutate(unbiased_rank = sig_terms_array[match(rownames(.), rownames(sig_terms_array)), "rank"], mean_cg = go_entrez2[match(rownames(.), go_entrez2$go), "mean_cg"]) %>% mutate(rank_change = unbiased_rank - rank) %>%
  ggplot(aes(x = log10(mean_cg), y = rank_change)) +
  geom_point()
```
- ok I don't think that's what i want that to look like
- maybe it would be better with all results

```{r}
sig_terms_array_biased %>% mutate(unbiased_rank = sig_terms_array[match(rownames(.), rownames(sig_terms_array)), "rank"], mean_cg = go_entrez2[match(rownames(.), go_entrez2$go), "mean_cg"]) %>% mutate(rank_change = rank - unbiased_rank) %>%
  ggplot(aes(x = mean_cg, y = rank_change)) +
  geom_point()

sig_terms_array_biased %>% mutate(unbiased_rank = sig_terms_array[match(rownames(.), rownames(sig_terms_array)), "rank"], mean_cg = go_entrez2[match(rownames(.), go_entrez2$go), "mean_cg"]) %>% mutate(rank_change = rank - unbiased_rank) %>%
  ggplot(aes(x = log10(mean_cg), y = rank_change)) +
  geom_point()
```


```{r}
test_b_num_unfilt
```

```{r}
test_b$rank <- 1:nrow(test_b)
test_b_null$rank <- 1:nrow(test_b_null)
```
```{r}
test_b
test_b_null
```
```{r}

```

```{r}
test_b %>% mutate(rank_null = test_b_null[match(rownames(.), rownames(test_b_null)), "rank"], mean_cg_bin = go_entrez2[match(rownames(.), go_entrez2$go), "mean_cg_bin"]) %>% 
  mutate(rank_change = rank_null - rank) %>%
  ggplot(aes(x = mean_cg_bin, y = rank_change)) +
  geom_point(alpha=0.3)

test_b %>% mutate(rank_null = test_b_null[match(rownames(.), rownames(test_b_null)), "rank"], mean_cg_bin = go_entrez2[match(rownames(.), go_entrez2$go), "mean_cg_bin"]) %>% 
  mutate(rank_change = rank_null - rank) %>%
  ggplot(aes(x = log10(mean_cg_bin), y = rank_change)) +
  geom_point(alpha=0.3)
```

- when on log scale - not what we want it to look like
```{r}
test_b %>% filter(FDR < 0.01)
test_b_null %>% filter(FDR < 0.01)
```
- well that's less than it was the first time
```{r}
test_b_num_dmr$rank <- 1:nrow(test_b_num_dmr)
test_b_num_dmr_null$rank <- 1:nrow(test_b_num_dmr_null)
```
```{r}
go_entrez2
```
```{r}
test_b_num_dmr %>% mutate(rank_null = test_b_num_dmr_null[match(rownames(.), rownames(test_b_num_dmr_null)), "rank"], mean_cg = go_entrez2[match(rownames(.), go_entrez2$go), "mean_cg"]) %>% mutate(rank_change = rank_null - rank)
```

```{r}
test_b_num_dmr %>% mutate(rank_null = test_b_num_dmr_null[match(rownames(.), rownames(test_b_num_dmr_null)), "rank"], mean_cg = go_entrez2[match(rownames(.), go_entrez2$go), "mean_cg"]) %>% 
  mutate(rank_change = rank_null - rank) %>%
  ggplot(aes(x = mean_cg, y = rank_change)) +
  geom_point(alpha=0.3)

test_b_num_dmr %>% mutate(rank_null = test_b_num_dmr_null[match(rownames(.), rownames(test_b_num_dmr_null)), "rank"], mean_cg = go_entrez2[match(rownames(.), go_entrez2$go), "mean_cg"]) %>% 
  mutate(rank_change = rank_null - rank) %>%
  ggplot(aes(x = log10(mean_cg), y = rank_change)) +
  geom_point(alpha=0.3)
```
